==> 01_Compute and display date/02_ComputeScaledData_CorrelationData.ipynb <==
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
%matplotlib inline
import seaborn as sns
import os

data_normal = pd.read_excel('All_Data.xlsx',sheet_name='data_normal')
scaled_data_normal = data_normal.copy()

list_column_data = data_normal.columns
columns_toBeScale = ["Iabc_generator_ContinueComponent_PhaseA", "Iabc_generator_ContinueComponent_PhaseB", "Iabc_generator_ContinueComponent_PhaseC", "Iabc_generator_EffectiveComponent_PhaseA", "Iabc_generator_EffectiveComponent_PhaseB", "Iabc_generator_EffectiveComponent_PhaseC", "Iabc_generator_FirstArmonic_PhaseA_Value", "Iabc_generator_FirstArmonic_PhaseB_Value", "Iabc_generator_FirstArmonic_PhaseC_Value", "Iabc_generator_DeformingComponent_PhaseA", "Iabc_generator_DeformingComponent_PhaseB", "Iabc_generator_DeformingComponent_PhaseC", "Iabc_generator_CoefficientDistorsion_PhaseA", "Iabc_generator_CoefficientDistorsion_PhaseB", "Iabc_generator_CoefficientDistorsion_PhaseC", "Iabc_generator_CoefficientDistorsionContinue_PhaseA", "Iabc_generator_CoefficientDistorsionContinue_PhaseB", "Iabc_generator_CoefficientDistorsionContinue_PhaseC", "Iabc_generator_CoefficientDistorsionFirstArmonic_PhaseA", "Iabc_generator_CoefficientDistorsionFirstArmonic_PhaseB", "Iabc_generator_CoefficientDistorsionFirstArmonic_PhaseC", "Iabc_generator_DirectComponent_Value", "Iabc_generator_InverseComponent_Value", "Iabc_generator_ZeroComponent_Value", "Iabc_generator_CoefficeintNegativeAsimetry", "Iabc_generator_CoefficeintZeroAsimetry", 
"Vabc_generator_ContinueComponent_PhaseA", "Vabc_generator_ContinueComponent_PhaseB", "Vabc_generator_ContinueComponent_PhaseC", "Vabc_generator_EffectiveComponent_PhaseA", "Vabc_generator_EffectiveComponent_PhaseB", "Vabc_generator_EffectiveComponent_PhaseC", "Vabc_generator_FirstArmonic_PhaseA_Value", "Vabc_generator_FirstArmonic_PhaseB_Value", "Vabc_generator_FirstArmonic_PhaseC_Value", "Vabc_generator_DeformingComponent_PhaseA", "Vabc_generator_DeformingComponent_PhaseB", "Vabc_generator_DeformingComponent_PhaseC", "Vabc_generator_CoefficientDistorsion_PhaseA", "Vabc_generator_CoefficientDistorsion_PhaseB", "Vabc_generator_CoefficientDistorsion_PhaseC", "Vabc_generator_CoefficientDistorsionContinue_PhaseA", "Vabc_generator_CoefficientDistorsionContinue_PhaseB", "Vabc_generator_CoefficientDistorsionContinue_PhaseC", "Vabc_generator_CoefficientDistorsionFirstArmonic_PhaseA", "Vabc_generator_CoefficientDistorsionFirstArmonic_PhaseB", "Vabc_generator_CoefficientDistorsionFirstArmonic_PhaseC", "Vabc_generator_DirectComponent_Value", "Vabc_generator_InverseComponent_Value", "Vabc_generator_ZeroComponent_Value", "Vabc_generator_CoefficeintNegativeAsimetry", "Vabc_generator_CoefficeintZeroAsimetry", 
"Iabc_loc_defect_ContinueComponent_PhaseA", "Iabc_loc_defect_ContinueComponent_PhaseB", "Iabc_loc_defect_ContinueComponent_PhaseC", "Iabc_loc_defect_EffectiveComponent_PhaseA", "Iabc_loc_defect_EffectiveComponent_PhaseB", "Iabc_loc_defect_EffectiveComponent_PhaseC", "Iabc_loc_defect_FirstArmonic_PhaseA_Value", "Iabc_loc_defect_FirstArmonic_PhaseB_Value", "Iabc_loc_defect_FirstArmonic_PhaseC_Value", "Iabc_loc_defect_DeformingComponent_PhaseA", "Iabc_loc_defect_DeformingComponent_PhaseB", "Iabc_loc_defect_DeformingComponent_PhaseC", "Iabc_loc_defect_CoefficientDistorsion_PhaseA", "Iabc_loc_defect_CoefficientDistorsion_PhaseB", "Iabc_loc_defect_CoefficientDistorsion_PhaseC", "Iabc_loc_defect_CoefficientDistorsionContinue_PhaseA", "Iabc_loc_defect_CoefficientDistorsionContinue_PhaseB", "Iabc_loc_defect_CoefficientDistorsionContinue_PhaseC", "Iabc_loc_defect_CoefficientDistorsionFirstArmonic_PhaseA", "Iabc_loc_defect_CoefficientDistorsionFirstArmonic_PhaseB", "Iabc_loc_defect_CoefficientDistorsionFirstArmonic_PhaseC", "Iabc_loc_defect_DirectComponent_Value", "Iabc_loc_defect_InverseComponent_Value", "Iabc_loc_defect_ZeroComponent_Value", "Iabc_loc_defect_CoefficeintNegativeAsimetry", "Iabc_loc_defect_CoefficeintZeroAsimetry", 
"Vabc_loc_defect_ContinueComponent_PhaseA", "Vabc_loc_defect_ContinueComponent_PhaseB", "Vabc_loc_defect_ContinueComponent_PhaseC", "Vabc_loc_defect_EffectiveComponent_PhaseA", "Vabc_loc_defect_EffectiveComponent_PhaseB", "Vabc_loc_defect_EffectiveComponent_PhaseC", "Vabc_loc_defect_FirstArmonic_PhaseA_Value", "Vabc_loc_defect_FirstArmonic_PhaseB_Value", "Vabc_loc_defect_FirstArmonic_PhaseC_Value", "Vabc_loc_defect_DeformingComponent_PhaseA", "Vabc_loc_defect_DeformingComponent_PhaseB", "Vabc_loc_defect_DeformingComponent_PhaseC", "Vabc_loc_defect_CoefficientDistorsion_PhaseA", "Vabc_loc_defect_CoefficientDistorsion_PhaseB", "Vabc_loc_defect_CoefficientDistorsion_PhaseC", "Vabc_loc_defect_CoefficientDistorsionContinue_PhaseA", "Vabc_loc_defect_CoefficientDistorsionContinue_PhaseB", "Vabc_loc_defect_CoefficientDistorsionContinue_PhaseC", "Vabc_loc_defect_CoefficientDistorsionFirstArmonic_PhaseA", "Vabc_loc_defect_CoefficientDistorsionFirstArmonic_PhaseB", "Vabc_loc_defect_CoefficientDistorsionFirstArmonic_PhaseC", "Vabc_loc_defect_DirectComponent_Value", "Vabc_loc_defect_InverseComponent_Value", "Vabc_loc_defect_ZeroComponent_Value", "Vabc_loc_defect_CoefficeintNegativeAsimetry", "Vabc_loc_defect_CoefficeintZeroAsimetry", 
"Iabc_consumator_ContinueComponent_PhaseA", "Iabc_consumator_ContinueComponent_PhaseB", "Iabc_consumator_ContinueComponent_PhaseC", "Iabc_consumator_EffectiveComponent_PhaseA", "Iabc_consumator_EffectiveComponent_PhaseB", "Iabc_consumator_EffectiveComponent_PhaseC", "Iabc_consumator_FirstArmonic_PhaseA_Value", "Iabc_consumator_FirstArmonic_PhaseB_Value", "Iabc_consumator_FirstArmonic_PhaseC_Value", "Iabc_consumator_DeformingComponent_PhaseA", "Iabc_consumator_DeformingComponent_PhaseB", "Iabc_consumator_DeformingComponent_PhaseC", "Iabc_consumator_CoefficientDistorsion_PhaseA", "Iabc_consumator_CoefficientDistorsion_PhaseB", "Iabc_consumator_CoefficientDistorsion_PhaseC", "Iabc_consumator_CoefficientDistorsionContinue_PhaseA", "Iabc_consumator_CoefficientDistorsionContinue_PhaseB", "Iabc_consumator_CoefficientDistorsionContinue_PhaseC", "Iabc_consumator_CoefficientDistorsionFirstArmonic_PhaseA", "Iabc_consumator_CoefficientDistorsionFirstArmonic_PhaseB", "Iabc_consumator_CoefficientDistorsionFirstArmonic_PhaseC", "Iabc_consumator_DirectComponent_Value", "Iabc_consumator_InverseComponent_Value", "Iabc_consumator_ZeroComponent_Value", "Iabc_consumator_CoefficeintNegativeAsimetry", "Iabc_consumator_CoefficeintZeroAsimetry", 
"Vabc_consumator_ContinueComponent_PhaseA", "Vabc_consumator_ContinueComponent_PhaseB", "Vabc_consumator_ContinueComponent_PhaseC", "Vabc_consumator_EffectiveComponent_PhaseA", "Vabc_consumator_EffectiveComponent_PhaseB", "Vabc_consumator_EffectiveComponent_PhaseC", "Vabc_consumator_FirstArmonic_PhaseA_Value", "Vabc_consumator_FirstArmonic_PhaseB_Value", "Vabc_consumator_FirstArmonic_PhaseC_Value", "Vabc_consumator_DeformingComponent_PhaseA", "Vabc_consumator_DeformingComponent_PhaseB", "Vabc_consumator_DeformingComponent_PhaseC", "Vabc_consumator_CoefficientDistorsion_PhaseA", "Vabc_consumator_CoefficientDistorsion_PhaseB", "Vabc_consumator_CoefficientDistorsion_PhaseC", "Vabc_consumator_CoefficientDistorsionContinue_PhaseA", "Vabc_consumator_CoefficientDistorsionContinue_PhaseB", "Vabc_consumator_CoefficientDistorsionContinue_PhaseC", "Vabc_consumator_CoefficientDistorsionFirstArmonic_PhaseA", "Vabc_consumator_CoefficientDistorsionFirstArmonic_PhaseB", "Vabc_consumator_CoefficientDistorsionFirstArmonic_PhaseC", "Vabc_consumator_DirectComponent_Value", "Vabc_consumator_InverseComponent_Value", "Vabc_consumator_ZeroComponent_Value", "Vabc_consumator_CoefficeintNegativeAsimetry", "Vabc_consumator_CoefficeintZeroAsimetry" ]
cases=set(data_normal["Type"])


cycleConst = 10
default_case = pd.concat([
    data_normal[(data_normal["Type"] == "MediumVoltage_ShortLine") & 
            (data_normal["cycle"] == cycleConst) & 
            (data_normal["phase"] == "A") &
            (data_normal["FaultLocation_FromGenerator"] == 5.0) &
            (data_normal["Nume_Grup"]=="tema 1_20kV_sarcina rezistiva" )  
           ], 
    data_normal[(data_normal["Type"] == "HighVoltage_ShortLine") & 
            (data_normal["cycle"] == cycleConst) & 
            (data_normal["phase"] == "A") &
            (data_normal["FaultLocation_FromGenerator"] == 5.0)
           ],
    data_normal[(data_normal["Type"] == "HighVoltage_LongLine500") & 
            (data_normal["cycle"] == cycleConst) & 
            (data_normal["phase"] == "A") &
            (data_normal["FaultLocation_FromGenerator"] == 250.0)
           ],
    data_normal[(data_normal["Type"] == "HighVoltage_LongLine300") & 
            (data_normal["cycle"] == cycleConst) & 
            (data_normal["phase"] == "A") &
            (data_normal["FaultLocation_FromGenerator"] == 150.0)
           ]
])
max_case = default_case.copy()


max_case["Iabc_generator_ContinueComponent_PhaseA"]=default_case[["Iabc_generator_ContinueComponent_PhaseA","Iabc_generator_ContinueComponent_PhaseB","Iabc_generator_ContinueComponent_PhaseC"]].max(axis=1)
max_case["Iabc_generator_ContinueComponent_PhaseB"]=default_case[["Iabc_generator_ContinueComponent_PhaseA","Iabc_generator_ContinueComponent_PhaseB","Iabc_generator_ContinueComponent_PhaseC"]].max(axis=1)
max_case["Iabc_generator_ContinueComponent_PhaseC"]=default_case[["Iabc_generator_ContinueComponent_PhaseA","Iabc_generator_ContinueComponent_PhaseB","Iabc_generator_ContinueComponent_PhaseC"]].max(axis=1)
max_case["Iabc_generator_EffectiveComponent_PhaseA"]=default_case[["Iabc_generator_EffectiveComponent_PhaseA","Iabc_generator_EffectiveComponent_PhaseB","Iabc_generator_EffectiveComponent_PhaseC"]].max(axis=1)
max_case["Iabc_generator_EffectiveComponent_PhaseB"]=default_case[["Iabc_generator_EffectiveComponent_PhaseA","Iabc_generator_EffectiveComponent_PhaseB","Iabc_generator_EffectiveComponent_PhaseC"]].max(axis=1)
max_case["Iabc_generator_EffectiveComponent_PhaseC"]=default_case[["Iabc_generator_EffectiveComponent_PhaseA","Iabc_generator_EffectiveComponent_PhaseB","Iabc_generator_EffectiveComponent_PhaseC"]].max(axis=1)
max_case["Iabc_generator_FirstArmonic_PhaseA_Value"]=default_case[["Iabc_generator_FirstArmonic_PhaseA_Value","Iabc_generator_FirstArmonic_PhaseB_Value","Iabc_generator_FirstArmonic_PhaseC_Value"]].max(axis=1)
max_case["Iabc_generator_FirstArmonic_PhaseB_Value"]=default_case[["Iabc_generator_FirstArmonic_PhaseA_Value","Iabc_generator_FirstArmonic_PhaseB_Value","Iabc_generator_FirstArmonic_PhaseC_Value"]].max(axis=1)
max_case["Iabc_generator_FirstArmonic_PhaseC_Value"]=default_case[["Iabc_generator_FirstArmonic_PhaseA_Value","Iabc_generator_FirstArmonic_PhaseB_Value","Iabc_generator_FirstArmonic_PhaseC_Value"]].max(axis=1)
max_case["Iabc_generator_DeformingComponent_PhaseA"]=default_case[["Iabc_generator_DeformingComponent_PhaseA","Iabc_generator_DeformingComponent_PhaseB","Iabc_generator_DeformingComponent_PhaseC"]].max(axis=1)
max_case["Iabc_generator_DeformingComponent_PhaseB"]=default_case[["Iabc_generator_DeformingComponent_PhaseA","Iabc_generator_DeformingComponent_PhaseB","Iabc_generator_DeformingComponent_PhaseC"]].max(axis=1)
max_case["Iabc_generator_DeformingComponent_PhaseC"]=default_case[["Iabc_generator_DeformingComponent_PhaseA","Iabc_generator_DeformingComponent_PhaseB","Iabc_generator_DeformingComponent_PhaseC"]].max(axis=1)
max_case["Iabc_generator_CoefficientDistorsion_PhaseA"]=default_case[["Iabc_generator_CoefficientDistorsion_PhaseA","Iabc_generator_CoefficientDistorsion_PhaseB","Iabc_generator_CoefficientDistorsion_PhaseC"]].max(axis=1)
max_case["Iabc_generator_CoefficientDistorsion_PhaseB"]=default_case[["Iabc_generator_CoefficientDistorsion_PhaseA","Iabc_generator_CoefficientDistorsion_PhaseB","Iabc_generator_CoefficientDistorsion_PhaseC"]].max(axis=1)
max_case["Iabc_generator_CoefficientDistorsion_PhaseC"]=default_case[["Iabc_generator_CoefficientDistorsion_PhaseA","Iabc_generator_CoefficientDistorsion_PhaseB","Iabc_generator_CoefficientDistorsion_PhaseC"]].max(axis=1)
max_case["Iabc_generator_CoefficientDistorsionContinue_PhaseA"]=default_case[["Iabc_generator_CoefficientDistorsionContinue_PhaseA","Iabc_generator_CoefficientDistorsionContinue_PhaseB","Iabc_generator_CoefficientDistorsionContinue_PhaseC"]].max(axis=1)
max_case["Iabc_generator_CoefficientDistorsionContinue_PhaseB"]=default_case[["Iabc_generator_CoefficientDistorsionContinue_PhaseA","Iabc_generator_CoefficientDistorsionContinue_PhaseB","Iabc_generator_CoefficientDistorsionContinue_PhaseC"]].max(axis=1)
max_case["Iabc_generator_CoefficientDistorsionContinue_PhaseC"]=default_case[["Iabc_generator_CoefficientDistorsionContinue_PhaseA","Iabc_generator_CoefficientDistorsionContinue_PhaseB","Iabc_generator_CoefficientDistorsionContinue_PhaseC"]].max(axis=1)
max_case["Iabc_generator_CoefficientDistorsionFirstArmonic_PhaseA"]=default_case[["Iabc_generator_CoefficientDistorsionFirstArmonic_PhaseA","Iabc_generator_CoefficientDistorsionFirstArmonic_PhaseB","Iabc_generator_CoefficientDistorsionFirstArmonic_PhaseC"]].max(axis=1)
max_case["Iabc_generator_CoefficientDistorsionFirstArmonic_PhaseB"]=default_case[["Iabc_generator_CoefficientDistorsionFirstArmonic_PhaseA","Iabc_generator_CoefficientDistorsionFirstArmonic_PhaseB","Iabc_generator_CoefficientDistorsionFirstArmonic_PhaseC"]].max(axis=1)
max_case["Iabc_generator_CoefficientDistorsionFirstArmonic_PhaseC"]=default_case[["Iabc_generator_CoefficientDistorsionFirstArmonic_PhaseA","Iabc_generator_CoefficientDistorsionFirstArmonic_PhaseB","Iabc_generator_CoefficientDistorsionFirstArmonic_PhaseC"]].max(axis=1)
max_case["Iabc_generator_DirectComponent_Value"]=default_case[["Iabc_generator_DirectComponent_Value","Iabc_generator_InverseComponent_Value","Iabc_generator_ZeroComponent_Value"]].max(axis=1)
max_case["Iabc_generator_InverseComponent_Value"]=default_case[["Iabc_generator_DirectComponent_Value","Iabc_generator_InverseComponent_Value","Iabc_generator_ZeroComponent_Value"]].max(axis=1)
max_case["Iabc_generator_ZeroComponent_Value"]=default_case[["Iabc_generator_DirectComponent_Value","Iabc_generator_InverseComponent_Value","Iabc_generator_ZeroComponent_Value"]].max(axis=1)

max_case["Vabc_generator_ContinueComponent_PhaseA"]=default_case[["Vabc_generator_ContinueComponent_PhaseA","Vabc_generator_ContinueComponent_PhaseB","Vabc_generator_ContinueComponent_PhaseC"]].max(axis=1)
max_case["Vabc_generator_ContinueComponent_PhaseB"]=default_case[["Vabc_generator_ContinueComponent_PhaseA","Vabc_generator_ContinueComponent_PhaseB","Vabc_generator_ContinueComponent_PhaseC"]].max(axis=1)
max_case["Vabc_generator_ContinueComponent_PhaseC"]=default_case[["Vabc_generator_ContinueComponent_PhaseA","Vabc_generator_ContinueComponent_PhaseB","Vabc_generator_ContinueComponent_PhaseC"]].max(axis=1)
max_case["Vabc_generator_EffectiveComponent_PhaseA"]=default_case[["Vabc_generator_EffectiveComponent_PhaseA","Vabc_generator_EffectiveComponent_PhaseB","Vabc_generator_EffectiveComponent_PhaseC"]].max(axis=1)
max_case["Vabc_generator_EffectiveComponent_PhaseB"]=default_case[["Vabc_generator_EffectiveComponent_PhaseA","Vabc_generator_EffectiveComponent_PhaseB","Vabc_generator_EffectiveComponent_PhaseC"]].max(axis=1)
max_case["Vabc_generator_EffectiveComponent_PhaseC"]=default_case[["Vabc_generator_EffectiveComponent_PhaseA","Vabc_generator_EffectiveComponent_PhaseB","Vabc_generator_EffectiveComponent_PhaseC"]].max(axis=1)
max_case["Vabc_generator_FirstArmonic_PhaseA_Value"]=default_case[["Vabc_generator_FirstArmonic_PhaseA_Value","Vabc_generator_FirstArmonic_PhaseB_Value","Vabc_generator_FirstArmonic_PhaseC_Value"]].max(axis=1)
max_case["Vabc_generator_FirstArmonic_PhaseB_Value"]=default_case[["Vabc_generator_FirstArmonic_PhaseA_Value","Vabc_generator_FirstArmonic_PhaseB_Value","Vabc_generator_FirstArmonic_PhaseC_Value"]].max(axis=1)
max_case["Vabc_generator_FirstArmonic_PhaseC_Value"]=default_case[["Vabc_generator_FirstArmonic_PhaseA_Value","Vabc_generator_FirstArmonic_PhaseB_Value","Vabc_generator_FirstArmonic_PhaseC_Value"]].max(axis=1)
max_case["Vabc_generator_DeformingComponent_PhaseA"]=default_case[["Vabc_generator_DeformingComponent_PhaseA","Vabc_generator_DeformingComponent_PhaseB","Vabc_generator_DeformingComponent_PhaseC"]].max(axis=1)
max_case["Vabc_generator_DeformingComponent_PhaseB"]=default_case[["Vabc_generator_DeformingComponent_PhaseA","Vabc_generator_DeformingComponent_PhaseB","Vabc_generator_DeformingComponent_PhaseC"]].max(axis=1)
max_case["Vabc_generator_DeformingComponent_PhaseC"]=default_case[["Vabc_generator_DeformingComponent_PhaseA","Vabc_generator_DeformingComponent_PhaseB","Vabc_generator_DeformingComponent_PhaseC"]].max(axis=1)
max_case["Vabc_generator_CoefficientDistorsion_PhaseA"]=default_case[["Vabc_generator_CoefficientDistorsion_PhaseA","Vabc_generator_CoefficientDistorsion_PhaseB","Vabc_generator_CoefficientDistorsion_PhaseC"]].max(axis=1)
max_case["Vabc_generator_CoefficientDistorsion_PhaseB"]=default_case[["Vabc_generator_CoefficientDistorsion_PhaseA","Vabc_generator_CoefficientDistorsion_PhaseB","Vabc_generator_CoefficientDistorsion_PhaseC"]].max(axis=1)
max_case["Vabc_generator_CoefficientDistorsion_PhaseC"]=default_case[["Vabc_generator_CoefficientDistorsion_PhaseA","Vabc_generator_CoefficientDistorsion_PhaseB","Vabc_generator_CoefficientDistorsion_PhaseC"]].max(axis=1)
max_case["Vabc_generator_CoefficientDistorsionContinue_PhaseA"]=default_case[["Vabc_generator_CoefficientDistorsionContinue_PhaseA","Vabc_generator_CoefficientDistorsionContinue_PhaseB","Vabc_generator_CoefficientDistorsionContinue_PhaseC"]].max(axis=1)
max_case["Vabc_generator_CoefficientDistorsionContinue_PhaseB"]=default_case[["Vabc_generator_CoefficientDistorsionContinue_PhaseA","Vabc_generator_CoefficientDistorsionContinue_PhaseB","Vabc_generator_CoefficientDistorsionContinue_PhaseC"]].max(axis=1)
max_case["Vabc_generator_CoefficientDistorsionContinue_PhaseC"]=default_case[["Vabc_generator_CoefficientDistorsionContinue_PhaseA","Vabc_generator_CoefficientDistorsionContinue_PhaseB","Vabc_generator_CoefficientDistorsionContinue_PhaseC"]].max(axis=1)
max_case["Vabc_generator_CoefficientDistorsionFirstArmonic_PhaseA"]=default_case[["Vabc_generator_CoefficientDistorsionFirstArmonic_PhaseA","Vabc_generator_CoefficientDistorsionFirstArmonic_PhaseB","Vabc_generator_CoefficientDistorsionFirstArmonic_PhaseC"]].max(axis=1)
max_case["Vabc_generator_CoefficientDistorsionFirstArmonic_PhaseB"]=default_case[["Vabc_generator_CoefficientDistorsionFirstArmonic_PhaseA","Vabc_generator_CoefficientDistorsionFirstArmonic_PhaseB","Vabc_generator_CoefficientDistorsionFirstArmonic_PhaseC"]].max(axis=1)
max_case["Vabc_generator_CoefficientDistorsionFirstArmonic_PhaseC"]=default_case[["Vabc_generator_CoefficientDistorsionFirstArmonic_PhaseA","Vabc_generator_CoefficientDistorsionFirstArmonic_PhaseB","Vabc_generator_CoefficientDistorsionFirstArmonic_PhaseC"]].max(axis=1)
max_case["Vabc_generator_DirectComponent_Value"]=default_case[["Vabc_generator_DirectComponent_Value","Vabc_generator_InverseComponent_Value","Vabc_generator_ZeroComponent_Value"]].max(axis=1)
max_case["Vabc_generator_InverseComponent_Value"]=default_case[["Vabc_generator_DirectComponent_Value","Vabc_generator_InverseComponent_Value","Vabc_generator_ZeroComponent_Value"]].max(axis=1)
max_case["Vabc_generator_ZeroComponent_Value"]=default_case[["Vabc_generator_DirectComponent_Value","Vabc_generator_InverseComponent_Value","Vabc_generator_ZeroComponent_Value"]].max(axis=1)

max_case["Iabc_loc_defect_ContinueComponent_PhaseA"]=default_case[["Iabc_loc_defect_ContinueComponent_PhaseA","Iabc_loc_defect_ContinueComponent_PhaseB","Iabc_loc_defect_ContinueComponent_PhaseC"]].max(axis=1)
max_case["Iabc_loc_defect_ContinueComponent_PhaseB"]=default_case[["Iabc_loc_defect_ContinueComponent_PhaseA","Iabc_loc_defect_ContinueComponent_PhaseB","Iabc_loc_defect_ContinueComponent_PhaseC"]].max(axis=1)
max_case["Iabc_loc_defect_ContinueComponent_PhaseC"]=default_case[["Iabc_loc_defect_ContinueComponent_PhaseA","Iabc_loc_defect_ContinueComponent_PhaseB","Iabc_loc_defect_ContinueComponent_PhaseC"]].max(axis=1)
max_case["Iabc_loc_defect_EffectiveComponent_PhaseA"]=default_case[["Iabc_loc_defect_EffectiveComponent_PhaseA","Iabc_loc_defect_EffectiveComponent_PhaseB","Iabc_loc_defect_EffectiveComponent_PhaseC"]].max(axis=1)
max_case["Iabc_loc_defect_EffectiveComponent_PhaseB"]=default_case[["Iabc_loc_defect_EffectiveComponent_PhaseA","Iabc_loc_defect_EffectiveComponent_PhaseB","Iabc_loc_defect_EffectiveComponent_PhaseC"]].max(axis=1)
max_case["Iabc_loc_defect_EffectiveComponent_PhaseC"]=default_case[["Iabc_loc_defect_EffectiveComponent_PhaseA","Iabc_loc_defect_EffectiveComponent_PhaseB","Iabc_loc_defect_EffectiveComponent_PhaseC"]].max(axis=1)
max_case["Iabc_loc_defect_FirstArmonic_PhaseA_Value"]=default_case[["Iabc_loc_defect_FirstArmonic_PhaseA_Value","Iabc_loc_defect_FirstArmonic_PhaseB_Value","Iabc_loc_defect_FirstArmonic_PhaseC_Value"]].max(axis=1)
max_case["Iabc_loc_defect_FirstArmonic_PhaseB_Value"]=default_case[["Iabc_loc_defect_FirstArmonic_PhaseA_Value","Iabc_loc_defect_FirstArmonic_PhaseB_Value","Iabc_loc_defect_FirstArmonic_PhaseC_Value"]].max(axis=1)
max_case["Iabc_loc_defect_FirstArmonic_PhaseC_Value"]=default_case[["Iabc_loc_defect_FirstArmonic_PhaseA_Value","Iabc_loc_defect_FirstArmonic_PhaseB_Value","Iabc_loc_defect_FirstArmonic_PhaseC_Value"]].max(axis=1)
max_case["Iabc_loc_defect_DeformingComponent_PhaseA"]=default_case[["Iabc_loc_defect_DeformingComponent_PhaseA","Iabc_loc_defect_DeformingComponent_PhaseB","Iabc_loc_defect_DeformingComponent_PhaseC"]].max(axis=1)
max_case["Iabc_loc_defect_DeformingComponent_PhaseB"]=default_case[["Iabc_loc_defect_DeformingComponent_PhaseA","Iabc_loc_defect_DeformingComponent_PhaseB","Iabc_loc_defect_DeformingComponent_PhaseC"]].max(axis=1)
max_case["Iabc_loc_defect_DeformingComponent_PhaseC"]=default_case[["Iabc_loc_defect_DeformingComponent_PhaseA","Iabc_loc_defect_DeformingComponent_PhaseB","Iabc_loc_defect_DeformingComponent_PhaseC"]].max(axis=1)
max_case["Iabc_loc_defect_CoefficientDistorsion_PhaseA"]=default_case[["Iabc_loc_defect_CoefficientDistorsion_PhaseA","Iabc_loc_defect_CoefficientDistorsion_PhaseB","Iabc_loc_defect_CoefficientDistorsion_PhaseC"]].max(axis=1)
max_case["Iabc_loc_defect_CoefficientDistorsion_PhaseB"]=default_case[["Iabc_loc_defect_CoefficientDistorsion_PhaseA","Iabc_loc_defect_CoefficientDistorsion_PhaseB","Iabc_loc_defect_CoefficientDistorsion_PhaseC"]].max(axis=1)
max_case["Iabc_loc_defect_CoefficientDistorsion_PhaseC"]=default_case[["Iabc_loc_defect_CoefficientDistorsion_PhaseA","Iabc_loc_defect_CoefficientDistorsion_PhaseB","Iabc_loc_defect_CoefficientDistorsion_PhaseC"]].max(axis=1)
max_case["Iabc_loc_defect_CoefficientDistorsionContinue_PhaseA"]=default_case[["Iabc_loc_defect_CoefficientDistorsionContinue_PhaseA","Iabc_loc_defect_CoefficientDistorsionContinue_PhaseB","Iabc_loc_defect_CoefficientDistorsionContinue_PhaseC"]].max(axis=1)
max_case["Iabc_loc_defect_CoefficientDistorsionContinue_PhaseB"]=default_case[["Iabc_loc_defect_CoefficientDistorsionContinue_PhaseA","Iabc_loc_defect_CoefficientDistorsionContinue_PhaseB","Iabc_loc_defect_CoefficientDistorsionContinue_PhaseC"]].max(axis=1)
max_case["Iabc_loc_defect_CoefficientDistorsionContinue_PhaseC"]=default_case[["Iabc_loc_defect_CoefficientDistorsionContinue_PhaseA","Iabc_loc_defect_CoefficientDistorsionContinue_PhaseB","Iabc_loc_defect_CoefficientDistorsionContinue_PhaseC"]].max(axis=1)
max_case["Iabc_loc_defect_CoefficientDistorsionFirstArmonic_PhaseA"]=default_case[["Iabc_loc_defect_CoefficientDistorsionFirstArmonic_PhaseA","Iabc_loc_defect_CoefficientDistorsionFirstArmonic_PhaseB","Iabc_loc_defect_CoefficientDistorsionFirstArmonic_PhaseC"]].max(axis=1)
max_case["Iabc_loc_defect_CoefficientDistorsionFirstArmonic_PhaseB"]=default_case[["Iabc_loc_defect_CoefficientDistorsionFirstArmonic_PhaseA","Iabc_loc_defect_CoefficientDistorsionFirstArmonic_PhaseB","Iabc_loc_defect_CoefficientDistorsionFirstArmonic_PhaseC"]].max(axis=1)
max_case["Iabc_loc_defect_CoefficientDistorsionFirstArmonic_PhaseC"]=default_case[["Iabc_loc_defect_CoefficientDistorsionFirstArmonic_PhaseA","Iabc_loc_defect_CoefficientDistorsionFirstArmonic_PhaseB","Iabc_loc_defect_CoefficientDistorsionFirstArmonic_PhaseC"]].max(axis=1)
max_case["Iabc_loc_defect_DirectComponent_Value"]=default_case[["Iabc_loc_defect_DirectComponent_Value","Iabc_loc_defect_InverseComponent_Value","Iabc_loc_defect_ZeroComponent_Value"]].max(axis=1)
max_case["Iabc_loc_defect_InverseComponent_Value"]=default_case[["Iabc_loc_defect_DirectComponent_Value","Iabc_loc_defect_InverseComponent_Value","Iabc_loc_defect_ZeroComponent_Value"]].max(axis=1)
max_case["Iabc_loc_defect_ZeroComponent_Value"]=default_case[["Iabc_loc_defect_DirectComponent_Value","Iabc_loc_defect_InverseComponent_Value","Iabc_loc_defect_ZeroComponent_Value"]].max(axis=1)

max_case["Vabc_loc_defect_ContinueComponent_PhaseA"]=default_case[["Vabc_loc_defect_ContinueComponent_PhaseA","Vabc_loc_defect_ContinueComponent_PhaseB","Vabc_loc_defect_ContinueComponent_PhaseC"]].max(axis=1)
max_case["Vabc_loc_defect_ContinueComponent_PhaseB"]=default_case[["Vabc_loc_defect_ContinueComponent_PhaseA","Vabc_loc_defect_ContinueComponent_PhaseB","Vabc_loc_defect_ContinueComponent_PhaseC"]].max(axis=1)
max_case["Vabc_loc_defect_ContinueComponent_PhaseC"]=default_case[["Vabc_loc_defect_ContinueComponent_PhaseA","Vabc_loc_defect_ContinueComponent_PhaseB","Vabc_loc_defect_ContinueComponent_PhaseC"]].max(axis=1)
max_case["Vabc_loc_defect_EffectiveComponent_PhaseA"]=default_case[["Vabc_loc_defect_EffectiveComponent_PhaseA","Vabc_loc_defect_EffectiveComponent_PhaseB","Vabc_loc_defect_EffectiveComponent_PhaseC"]].max(axis=1)
max_case["Vabc_loc_defect_EffectiveComponent_PhaseB"]=default_case[["Vabc_loc_defect_EffectiveComponent_PhaseA","Vabc_loc_defect_EffectiveComponent_PhaseB","Vabc_loc_defect_EffectiveComponent_PhaseC"]].max(axis=1)
max_case["Vabc_loc_defect_EffectiveComponent_PhaseC"]=default_case[["Vabc_loc_defect_EffectiveComponent_PhaseA","Vabc_loc_defect_EffectiveComponent_PhaseB","Vabc_loc_defect_EffectiveComponent_PhaseC"]].max(axis=1)
max_case["Vabc_loc_defect_FirstArmonic_PhaseA_Value"]=default_case[["Vabc_loc_defect_FirstArmonic_PhaseA_Value","Vabc_loc_defect_FirstArmonic_PhaseB_Value","Vabc_loc_defect_FirstArmonic_PhaseC_Value"]].max(axis=1)
max_case["Vabc_loc_defect_FirstArmonic_PhaseB_Value"]=default_case[["Vabc_loc_defect_FirstArmonic_PhaseA_Value","Vabc_loc_defect_FirstArmonic_PhaseB_Value","Vabc_loc_defect_FirstArmonic_PhaseC_Value"]].max(axis=1)
max_case["Vabc_loc_defect_FirstArmonic_PhaseC_Value"]=default_case[["Vabc_loc_defect_FirstArmonic_PhaseA_Value","Vabc_loc_defect_FirstArmonic_PhaseB_Value","Vabc_loc_defect_FirstArmonic_PhaseC_Value"]].max(axis=1)
max_case["Vabc_loc_defect_DeformingComponent_PhaseA"]=default_case[["Vabc_loc_defect_DeformingComponent_PhaseA","Vabc_loc_defect_DeformingComponent_PhaseB","Vabc_loc_defect_DeformingComponent_PhaseC"]].max(axis=1)
max_case["Vabc_loc_defect_DeformingComponent_PhaseB"]=default_case[["Vabc_loc_defect_DeformingComponent_PhaseA","Vabc_loc_defect_DeformingComponent_PhaseB","Vabc_loc_defect_DeformingComponent_PhaseC"]].max(axis=1)
max_case["Vabc_loc_defect_DeformingComponent_PhaseC"]=default_case[["Vabc_loc_defect_DeformingComponent_PhaseA","Vabc_loc_defect_DeformingComponent_PhaseB","Vabc_loc_defect_DeformingComponent_PhaseC"]].max(axis=1)
max_case["Vabc_loc_defect_CoefficientDistorsion_PhaseA"]=default_case[["Vabc_loc_defect_CoefficientDistorsion_PhaseA","Vabc_loc_defect_CoefficientDistorsion_PhaseB","Vabc_loc_defect_CoefficientDistorsion_PhaseC"]].max(axis=1)
max_case["Vabc_loc_defect_CoefficientDistorsion_PhaseB"]=default_case[["Vabc_loc_defect_CoefficientDistorsion_PhaseA","Vabc_loc_defect_CoefficientDistorsion_PhaseB","Vabc_loc_defect_CoefficientDistorsion_PhaseC"]].max(axis=1)
max_case["Vabc_loc_defect_CoefficientDistorsion_PhaseC"]=default_case[["Vabc_loc_defect_CoefficientDistorsion_PhaseA","Vabc_loc_defect_CoefficientDistorsion_PhaseB","Vabc_loc_defect_CoefficientDistorsion_PhaseC"]].max(axis=1)
max_case["Vabc_loc_defect_CoefficientDistorsionContinue_PhaseA"]=default_case[["Vabc_loc_defect_CoefficientDistorsionContinue_PhaseA","Vabc_loc_defect_CoefficientDistorsionContinue_PhaseB","Vabc_loc_defect_CoefficientDistorsionContinue_PhaseC"]].max(axis=1)
max_case["Vabc_loc_defect_CoefficientDistorsionContinue_PhaseB"]=default_case[["Vabc_loc_defect_CoefficientDistorsionContinue_PhaseA","Vabc_loc_defect_CoefficientDistorsionContinue_PhaseB","Vabc_loc_defect_CoefficientDistorsionContinue_PhaseC"]].max(axis=1)
max_case["Vabc_loc_defect_CoefficientDistorsionContinue_PhaseC"]=default_case[["Vabc_loc_defect_CoefficientDistorsionContinue_PhaseA","Vabc_loc_defect_CoefficientDistorsionContinue_PhaseB","Vabc_loc_defect_CoefficientDistorsionContinue_PhaseC"]].max(axis=1)
max_case["Vabc_loc_defect_CoefficientDistorsionFirstArmonic_PhaseA"]=default_case[["Vabc_loc_defect_CoefficientDistorsionFirstArmonic_PhaseA","Vabc_loc_defect_CoefficientDistorsionFirstArmonic_PhaseB","Vabc_loc_defect_CoefficientDistorsionFirstArmonic_PhaseC"]].max(axis=1)
max_case["Vabc_loc_defect_CoefficientDistorsionFirstArmonic_PhaseB"]=default_case[["Vabc_loc_defect_CoefficientDistorsionFirstArmonic_PhaseA","Vabc_loc_defect_CoefficientDistorsionFirstArmonic_PhaseB","Vabc_loc_defect_CoefficientDistorsionFirstArmonic_PhaseC"]].max(axis=1)
max_case["Vabc_loc_defect_CoefficientDistorsionFirstArmonic_PhaseC"]=default_case[["Vabc_loc_defect_CoefficientDistorsionFirstArmonic_PhaseA","Vabc_loc_defect_CoefficientDistorsionFirstArmonic_PhaseB","Vabc_loc_defect_CoefficientDistorsionFirstArmonic_PhaseC"]].max(axis=1)
max_case["Vabc_loc_defect_DirectComponent_Value"]=default_case[["Vabc_loc_defect_DirectComponent_Value","Vabc_loc_defect_InverseComponent_Value","Vabc_loc_defect_ZeroComponent_Value"]].max(axis=1)
max_case["Vabc_loc_defect_InverseComponent_Value"]=default_case[["Vabc_loc_defect_DirectComponent_Value","Vabc_loc_defect_InverseComponent_Value","Vabc_loc_defect_ZeroComponent_Value"]].max(axis=1)
max_case["Vabc_loc_defect_ZeroComponent_Value"]=default_case[["Vabc_loc_defect_DirectComponent_Value","Vabc_loc_defect_InverseComponent_Value","Vabc_loc_defect_ZeroComponent_Value"]].max(axis=1)

max_case["Iabc_consumator_ContinueComponent_PhaseA"]=default_case[["Iabc_consumator_ContinueComponent_PhaseA","Iabc_consumator_ContinueComponent_PhaseB","Iabc_consumator_ContinueComponent_PhaseC"]].max(axis=1)
max_case["Iabc_consumator_ContinueComponent_PhaseB"]=default_case[["Iabc_consumator_ContinueComponent_PhaseA","Iabc_consumator_ContinueComponent_PhaseB","Iabc_consumator_ContinueComponent_PhaseC"]].max(axis=1)
max_case["Iabc_consumator_ContinueComponent_PhaseC"]=default_case[["Iabc_consumator_ContinueComponent_PhaseA","Iabc_consumator_ContinueComponent_PhaseB","Iabc_consumator_ContinueComponent_PhaseC"]].max(axis=1)
max_case["Iabc_consumator_EffectiveComponent_PhaseA"]=default_case[["Iabc_consumator_EffectiveComponent_PhaseA","Iabc_consumator_EffectiveComponent_PhaseB","Iabc_consumator_EffectiveComponent_PhaseC"]].max(axis=1)
max_case["Iabc_consumator_EffectiveComponent_PhaseB"]=default_case[["Iabc_consumator_EffectiveComponent_PhaseA","Iabc_consumator_EffectiveComponent_PhaseB","Iabc_consumator_EffectiveComponent_PhaseC"]].max(axis=1)
max_case["Iabc_consumator_EffectiveComponent_PhaseC"]=default_case[["Iabc_consumator_EffectiveComponent_PhaseA","Iabc_consumator_EffectiveComponent_PhaseB","Iabc_consumator_EffectiveComponent_PhaseC"]].max(axis=1)
max_case["Iabc_consumator_FirstArmonic_PhaseA_Value"]=default_case[["Iabc_consumator_FirstArmonic_PhaseA_Value","Iabc_consumator_FirstArmonic_PhaseB_Value","Iabc_consumator_FirstArmonic_PhaseC_Value"]].max(axis=1)
max_case["Iabc_consumator_FirstArmonic_PhaseB_Value"]=default_case[["Iabc_consumator_FirstArmonic_PhaseA_Value","Iabc_consumator_FirstArmonic_PhaseB_Value","Iabc_consumator_FirstArmonic_PhaseC_Value"]].max(axis=1)
max_case["Iabc_consumator_FirstArmonic_PhaseC_Value"]=default_case[["Iabc_consumator_FirstArmonic_PhaseA_Value","Iabc_consumator_FirstArmonic_PhaseB_Value","Iabc_consumator_FirstArmonic_PhaseC_Value"]].max(axis=1)
max_case["Iabc_consumator_DeformingComponent_PhaseA"]=default_case[["Iabc_consumator_DeformingComponent_PhaseA","Iabc_consumator_DeformingComponent_PhaseB","Iabc_consumator_DeformingComponent_PhaseC"]].max(axis=1)
max_case["Iabc_consumator_DeformingComponent_PhaseB"]=default_case[["Iabc_consumator_DeformingComponent_PhaseA","Iabc_consumator_DeformingComponent_PhaseB","Iabc_consumator_DeformingComponent_PhaseC"]].max(axis=1)
max_case["Iabc_consumator_DeformingComponent_PhaseC"]=default_case[["Iabc_consumator_DeformingComponent_PhaseA","Iabc_consumator_DeformingComponent_PhaseB","Iabc_consumator_DeformingComponent_PhaseC"]].max(axis=1)
max_case["Iabc_consumator_CoefficientDistorsion_PhaseA"]=default_case[["Iabc_consumator_CoefficientDistorsion_PhaseA","Iabc_consumator_CoefficientDistorsion_PhaseB","Iabc_consumator_CoefficientDistorsion_PhaseC"]].max(axis=1)
max_case["Iabc_consumator_CoefficientDistorsion_PhaseB"]=default_case[["Iabc_consumator_CoefficientDistorsion_PhaseA","Iabc_consumator_CoefficientDistorsion_PhaseB","Iabc_consumator_CoefficientDistorsion_PhaseC"]].max(axis=1)
max_case["Iabc_consumator_CoefficientDistorsion_PhaseC"]=default_case[["Iabc_consumator_CoefficientDistorsion_PhaseA","Iabc_consumator_CoefficientDistorsion_PhaseB","Iabc_consumator_CoefficientDistorsion_PhaseC"]].max(axis=1)
max_case["Iabc_consumator_CoefficientDistorsionContinue_PhaseA"]=default_case[["Iabc_consumator_CoefficientDistorsionContinue_PhaseA","Iabc_consumator_CoefficientDistorsionContinue_PhaseB","Iabc_consumator_CoefficientDistorsionContinue_PhaseC"]].max(axis=1)
max_case["Iabc_consumator_CoefficientDistorsionContinue_PhaseB"]=default_case[["Iabc_consumator_CoefficientDistorsionContinue_PhaseA","Iabc_consumator_CoefficientDistorsionContinue_PhaseB","Iabc_consumator_CoefficientDistorsionContinue_PhaseC"]].max(axis=1)
max_case["Iabc_consumator_CoefficientDistorsionContinue_PhaseC"]=default_case[["Iabc_consumator_CoefficientDistorsionContinue_PhaseA","Iabc_consumator_CoefficientDistorsionContinue_PhaseB","Iabc_consumator_CoefficientDistorsionContinue_PhaseC"]].max(axis=1)
max_case["Iabc_consumator_CoefficientDistorsionFirstArmonic_PhaseA"]=default_case[["Iabc_consumator_CoefficientDistorsionFirstArmonic_PhaseA","Iabc_consumator_CoefficientDistorsionFirstArmonic_PhaseB","Iabc_consumator_CoefficientDistorsionFirstArmonic_PhaseC"]].max(axis=1)
max_case["Iabc_consumator_CoefficientDistorsionFirstArmonic_PhaseB"]=default_case[["Iabc_consumator_CoefficientDistorsionFirstArmonic_PhaseA","Iabc_consumator_CoefficientDistorsionFirstArmonic_PhaseB","Iabc_consumator_CoefficientDistorsionFirstArmonic_PhaseC"]].max(axis=1)
max_case["Iabc_consumator_CoefficientDistorsionFirstArmonic_PhaseC"]=default_case[["Iabc_consumator_CoefficientDistorsionFirstArmonic_PhaseA","Iabc_consumator_CoefficientDistorsionFirstArmonic_PhaseB","Iabc_consumator_CoefficientDistorsionFirstArmonic_PhaseC"]].max(axis=1)
max_case["Iabc_consumator_DirectComponent_Value"]=default_case[["Iabc_consumator_DirectComponent_Value","Iabc_consumator_InverseComponent_Value","Iabc_consumator_ZeroComponent_Value"]].max(axis=1)
max_case["Iabc_consumator_InverseComponent_Value"]=default_case[["Iabc_consumator_DirectComponent_Value","Iabc_consumator_InverseComponent_Value","Iabc_consumator_ZeroComponent_Value"]].max(axis=1)
max_case["Iabc_consumator_ZeroComponent_Value"]=default_case[["Iabc_consumator_DirectComponent_Value","Iabc_consumator_InverseComponent_Value","Iabc_consumator_ZeroComponent_Value"]].max(axis=1)

max_case["Vabc_consumator_ContinueComponent_PhaseA"]=default_case[["Vabc_consumator_ContinueComponent_PhaseA","Vabc_consumator_ContinueComponent_PhaseB","Vabc_consumator_ContinueComponent_PhaseC"]].max(axis=1)
max_case["Vabc_consumator_ContinueComponent_PhaseB"]=default_case[["Vabc_consumator_ContinueComponent_PhaseA","Vabc_consumator_ContinueComponent_PhaseB","Vabc_consumator_ContinueComponent_PhaseC"]].max(axis=1)
max_case["Vabc_consumator_ContinueComponent_PhaseC"]=default_case[["Vabc_consumator_ContinueComponent_PhaseA","Vabc_consumator_ContinueComponent_PhaseB","Vabc_consumator_ContinueComponent_PhaseC"]].max(axis=1)
max_case["Vabc_consumator_EffectiveComponent_PhaseA"]=default_case[["Vabc_consumator_EffectiveComponent_PhaseA","Vabc_consumator_EffectiveComponent_PhaseB","Vabc_consumator_EffectiveComponent_PhaseC"]].max(axis=1)
max_case["Vabc_consumator_EffectiveComponent_PhaseB"]=default_case[["Vabc_consumator_EffectiveComponent_PhaseA","Vabc_consumator_EffectiveComponent_PhaseB","Vabc_consumator_EffectiveComponent_PhaseC"]].max(axis=1)
max_case["Vabc_consumator_EffectiveComponent_PhaseC"]=default_case[["Vabc_consumator_EffectiveComponent_PhaseA","Vabc_consumator_EffectiveComponent_PhaseB","Vabc_consumator_EffectiveComponent_PhaseC"]].max(axis=1)
max_case["Vabc_consumator_FirstArmonic_PhaseA_Value"]=default_case[["Vabc_consumator_FirstArmonic_PhaseA_Value","Vabc_consumator_FirstArmonic_PhaseB_Value","Vabc_consumator_FirstArmonic_PhaseC_Value"]].max(axis=1)
max_case["Vabc_consumator_FirstArmonic_PhaseB_Value"]=default_case[["Vabc_consumator_FirstArmonic_PhaseA_Value","Vabc_consumator_FirstArmonic_PhaseB_Value","Vabc_consumator_FirstArmonic_PhaseC_Value"]].max(axis=1)
max_case["Vabc_consumator_FirstArmonic_PhaseC_Value"]=default_case[["Vabc_consumator_FirstArmonic_PhaseA_Value","Vabc_consumator_FirstArmonic_PhaseB_Value","Vabc_consumator_FirstArmonic_PhaseC_Value"]].max(axis=1)
max_case["Vabc_consumator_DeformingComponent_PhaseA"]=default_case[["Vabc_consumator_DeformingComponent_PhaseA","Vabc_consumator_DeformingComponent_PhaseB","Vabc_consumator_DeformingComponent_PhaseC"]].max(axis=1)
max_case["Vabc_consumator_DeformingComponent_PhaseB"]=default_case[["Vabc_consumator_DeformingComponent_PhaseA","Vabc_consumator_DeformingComponent_PhaseB","Vabc_consumator_DeformingComponent_PhaseC"]].max(axis=1)
max_case["Vabc_consumator_DeformingComponent_PhaseC"]=default_case[["Vabc_consumator_DeformingComponent_PhaseA","Vabc_consumator_DeformingComponent_PhaseB","Vabc_consumator_DeformingComponent_PhaseC"]].max(axis=1)
max_case["Vabc_consumator_CoefficientDistorsion_PhaseA"]=default_case[["Vabc_consumator_CoefficientDistorsion_PhaseA","Vabc_consumator_CoefficientDistorsion_PhaseB","Vabc_consumator_CoefficientDistorsion_PhaseC"]].max(axis=1)
max_case["Vabc_consumator_CoefficientDistorsion_PhaseB"]=default_case[["Vabc_consumator_CoefficientDistorsion_PhaseA","Vabc_consumator_CoefficientDistorsion_PhaseB","Vabc_consumator_CoefficientDistorsion_PhaseC"]].max(axis=1)
max_case["Vabc_consumator_CoefficientDistorsion_PhaseC"]=default_case[["Vabc_consumator_CoefficientDistorsion_PhaseA","Vabc_consumator_CoefficientDistorsion_PhaseB","Vabc_consumator_CoefficientDistorsion_PhaseC"]].max(axis=1)
max_case["Vabc_consumator_CoefficientDistorsionContinue_PhaseA"]=default_case[["Vabc_consumator_CoefficientDistorsionContinue_PhaseA","Vabc_consumator_CoefficientDistorsionContinue_PhaseB","Vabc_consumator_CoefficientDistorsionContinue_PhaseC"]].max(axis=1)
max_case["Vabc_consumator_CoefficientDistorsionContinue_PhaseB"]=default_case[["Vabc_consumator_CoefficientDistorsionContinue_PhaseA","Vabc_consumator_CoefficientDistorsionContinue_PhaseB","Vabc_consumator_CoefficientDistorsionContinue_PhaseC"]].max(axis=1)
max_case["Vabc_consumator_CoefficientDistorsionContinue_PhaseC"]=default_case[["Vabc_consumator_CoefficientDistorsionContinue_PhaseA","Vabc_consumator_CoefficientDistorsionContinue_PhaseB","Vabc_consumator_CoefficientDistorsionContinue_PhaseC"]].max(axis=1)
max_case["Vabc_consumator_CoefficientDistorsionFirstArmonic_PhaseA"]=default_case[["Vabc_consumator_CoefficientDistorsionFirstArmonic_PhaseA","Vabc_consumator_CoefficientDistorsionFirstArmonic_PhaseB","Vabc_consumator_CoefficientDistorsionFirstArmonic_PhaseC"]].max(axis=1)
max_case["Vabc_consumator_CoefficientDistorsionFirstArmonic_PhaseB"]=default_case[["Vabc_consumator_CoefficientDistorsionFirstArmonic_PhaseA","Vabc_consumator_CoefficientDistorsionFirstArmonic_PhaseB","Vabc_consumator_CoefficientDistorsionFirstArmonic_PhaseC"]].max(axis=1)
max_case["Vabc_consumator_CoefficientDistorsionFirstArmonic_PhaseC"]=default_case[["Vabc_consumator_CoefficientDistorsionFirstArmonic_PhaseA","Vabc_consumator_CoefficientDistorsionFirstArmonic_PhaseB","Vabc_consumator_CoefficientDistorsionFirstArmonic_PhaseC"]].max(axis=1)
max_case["Vabc_consumator_DirectComponent_Value"]=default_case[["Vabc_consumator_DirectComponent_Value","Vabc_consumator_InverseComponent_Value","Vabc_consumator_ZeroComponent_Value"]].max(axis=1)
max_case["Vabc_consumator_InverseComponent_Value"]=default_case[["Vabc_consumator_DirectComponent_Value","Vabc_consumator_InverseComponent_Value","Vabc_consumator_ZeroComponent_Value"]].max(axis=1)
max_case["Vabc_consumator_ZeroComponent_Value"]=default_case[["Vabc_consumator_DirectComponent_Value","Vabc_consumator_InverseComponent_Value","Vabc_consumator_ZeroComponent_Value"]].max(axis=1)

for case in cases:
    data_normal_case = data_normal[(data_normal["Type"]==case)]
    max_single_case = max_case[(max_case["Type"]==case)]
    
    for column in columns_toBeScale:
        max_column=max_single_case[column]
        data_normal_case_column = data_normal_case[column]
        for i in data_normal_case_column.index.to_list():
            scaled_data_normal.at[i, column] = data_normal_case_column.at[i]/max_column
            scaled_data_normal.at[i, "FaultLocation_FromGenerator"]=data_normal_case.at[i, "FaultLocation_FromGenerator"] / data_normal_case.at[i, "TotalLength"]
            scaled_data_normal.at[i, "FaultLocation_ToConsumer"]=data_normal_case.at[i, "FaultLocation_ToConsumer"] / data_normal_case.at[i, "TotalLength"]

max_case.to_csv('Normal_Scaled_Factor.csv')

column_fault_machine = ["ID_Simulare", "obs_Simulare", "Nume_Grup", "Simulation_name", "Type", "phase", "FaultLocation_FromGenerator", "FaultLocation_ToConsumer", "TotalLength", "cycle", 
"Iabc_generator_ContinueComponent_PhaseA", "Iabc_generator_ContinueComponent_PhaseB", "Iabc_generator_ContinueComponent_PhaseC", "Iabc_generator_EffectiveComponent_PhaseA", "Iabc_generator_EffectiveComponent_PhaseB", "Iabc_generator_EffectiveComponent_PhaseC", "Iabc_generator_FirstArmonic_PhaseA_Value", "Iabc_generator_FirstArmonic_PhaseA_Angle", "Iabc_generator_FirstArmonic_PhaseB_Value", "Iabc_generator_FirstArmonic_PhaseB_Angle", "Iabc_generator_FirstArmonic_PhaseC_Value", "Iabc_generator_FirstArmonic_PhaseC_Angle", "Iabc_generator_DeformingComponent_PhaseA", "Iabc_generator_DeformingComponent_PhaseB", "Iabc_generator_DeformingComponent_PhaseC", "Iabc_generator_CoefficientDistorsion_PhaseA", "Iabc_generator_CoefficientDistorsion_PhaseB", "Iabc_generator_CoefficientDistorsion_PhaseC", "Iabc_generator_CoefficientDistorsionContinue_PhaseA", "Iabc_generator_CoefficientDistorsionContinue_PhaseB", "Iabc_generator_CoefficientDistorsionContinue_PhaseC", "Iabc_generator_CoefficientDistorsionFirstArmonic_PhaseA", "Iabc_generator_CoefficientDistorsionFirstArmonic_PhaseB", "Iabc_generator_CoefficientDistorsionFirstArmonic_PhaseC", "Iabc_generator_DirectComponent_Value", "Iabc_generator_DirectComponent_Angle", "Iabc_generator_InverseComponent_Value", "Iabc_generator_InverseComponent_Angle", "Iabc_generator_ZeroComponent_Value", "Iabc_generator_ZeroComponent_Angle", "Iabc_generator_CoefficeintNegativeAsimetry", "Iabc_generator_CoefficeintZeroAsimetry", 
"Vabc_generator_ContinueComponent_PhaseA", "Vabc_generator_ContinueComponent_PhaseB", "Vabc_generator_ContinueComponent_PhaseC", "Vabc_generator_EffectiveComponent_PhaseA", "Vabc_generator_EffectiveComponent_PhaseB", "Vabc_generator_EffectiveComponent_PhaseC", "Vabc_generator_FirstArmonic_PhaseA_Value", "Vabc_generator_FirstArmonic_PhaseA_Angle", "Vabc_generator_FirstArmonic_PhaseB_Value", "Vabc_generator_FirstArmonic_PhaseB_Angle", "Vabc_generator_FirstArmonic_PhaseC_Value", "Vabc_generator_FirstArmonic_PhaseC_Angle", "Vabc_generator_DeformingComponent_PhaseA", "Vabc_generator_DeformingComponent_PhaseB", "Vabc_generator_DeformingComponent_PhaseC", "Vabc_generator_CoefficientDistorsion_PhaseA", "Vabc_generator_CoefficientDistorsion_PhaseB", "Vabc_generator_CoefficientDistorsion_PhaseC", "Vabc_generator_CoefficientDistorsionContinue_PhaseA", "Vabc_generator_CoefficientDistorsionContinue_PhaseB", "Vabc_generator_CoefficientDistorsionContinue_PhaseC", "Vabc_generator_CoefficientDistorsionFirstArmonic_PhaseA", "Vabc_generator_CoefficientDistorsionFirstArmonic_PhaseB", "Vabc_generator_CoefficientDistorsionFirstArmonic_PhaseC", "Vabc_generator_DirectComponent_Value", "Vabc_generator_DirectComponent_Angle", "Vabc_generator_InverseComponent_Value", "Vabc_generator_InverseComponent_Angle", "Vabc_generator_ZeroComponent_Value", "Vabc_generator_ZeroComponent_Angle", "Vabc_generator_CoefficeintNegativeAsimetry", "Vabc_generator_CoefficeintZeroAsimetry", 
"Iabc_consumator_ContinueComponent_PhaseA", "Iabc_consumator_ContinueComponent_PhaseB", "Iabc_consumator_ContinueComponent_PhaseC", "Iabc_consumator_EffectiveComponent_PhaseA", "Iabc_consumator_EffectiveComponent_PhaseB", "Iabc_consumator_EffectiveComponent_PhaseC", "Iabc_consumator_FirstArmonic_PhaseA_Value", "Iabc_consumator_FirstArmonic_PhaseA_Angle", "Iabc_consumator_FirstArmonic_PhaseB_Value", "Iabc_consumator_FirstArmonic_PhaseB_Angle", "Iabc_consumator_FirstArmonic_PhaseC_Value", "Iabc_consumator_FirstArmonic_PhaseC_Angle", "Iabc_consumator_DeformingComponent_PhaseA", "Iabc_consumator_DeformingComponent_PhaseB", "Iabc_consumator_DeformingComponent_PhaseC", "Iabc_consumator_CoefficientDistorsion_PhaseA", "Iabc_consumator_CoefficientDistorsion_PhaseB", "Iabc_consumator_CoefficientDistorsion_PhaseC", "Iabc_consumator_CoefficientDistorsionContinue_PhaseA", "Iabc_consumator_CoefficientDistorsionContinue_PhaseB", "Iabc_consumator_CoefficientDistorsionContinue_PhaseC", "Iabc_consumator_CoefficientDistorsionFirstArmonic_PhaseA", "Iabc_consumator_CoefficientDistorsionFirstArmonic_PhaseB", "Iabc_consumator_CoefficientDistorsionFirstArmonic_PhaseC", "Iabc_consumator_DirectComponent_Value", "Iabc_consumator_DirectComponent_Angle", "Iabc_consumator_InverseComponent_Value", "Iabc_consumator_InverseComponent_Angle", "Iabc_consumator_ZeroComponent_Value", "Iabc_consumator_ZeroComponent_Angle", "Iabc_consumator_CoefficeintNegativeAsimetry", "Iabc_consumator_CoefficeintZeroAsimetry", 
"Vabc_consumator_ContinueComponent_PhaseA", "Vabc_consumator_ContinueComponent_PhaseB", "Vabc_consumator_ContinueComponent_PhaseC", "Vabc_consumator_EffectiveComponent_PhaseA", "Vabc_consumator_EffectiveComponent_PhaseB", "Vabc_consumator_EffectiveComponent_PhaseC", "Vabc_consumator_FirstArmonic_PhaseA_Value", "Vabc_consumator_FirstArmonic_PhaseA_Angle", "Vabc_consumator_FirstArmonic_PhaseB_Value", "Vabc_consumator_FirstArmonic_PhaseB_Angle", "Vabc_consumator_FirstArmonic_PhaseC_Value", "Vabc_consumator_FirstArmonic_PhaseC_Angle", "Vabc_consumator_DeformingComponent_PhaseA", "Vabc_consumator_DeformingComponent_PhaseB", "Vabc_consumator_DeformingComponent_PhaseC", "Vabc_consumator_CoefficientDistorsion_PhaseA", "Vabc_consumator_CoefficientDistorsion_PhaseB", "Vabc_consumator_CoefficientDistorsion_PhaseC", "Vabc_consumator_CoefficientDistorsionContinue_PhaseA", "Vabc_consumator_CoefficientDistorsionContinue_PhaseB", "Vabc_consumator_CoefficientDistorsionContinue_PhaseC", "Vabc_consumator_CoefficientDistorsionFirstArmonic_PhaseA", "Vabc_consumator_CoefficientDistorsionFirstArmonic_PhaseB", "Vabc_consumator_CoefficientDistorsionFirstArmonic_PhaseC", "Vabc_consumator_DirectComponent_Value", "Vabc_consumator_DirectComponent_Angle", "Vabc_consumator_InverseComponent_Value", "Vabc_consumator_InverseComponent_Angle", "Vabc_consumator_ZeroComponent_Value", "Vabc_consumator_ZeroComponent_Angle", "Vabc_consumator_CoefficeintNegativeAsimetry", "Vabc_consumator_CoefficeintZeroAsimetry", ]

data_normal_fault=data_normal[column_fault_machine].copy()
scaled_data_normal_fault=scaled_data_normal[column_fault_machine].copy()

data_normal.to_csv('Save_Normal_data.csv')
scaled_data_normal.to_csv('Save_Normal_data_SCALED.csv')
data_normal_fault.to_csv('Save_Fault_data.csv')
scaled_data_normal_fault.to_csv('Save_Fault_data_SCALED.csv')

columns_to_analyze = list(scaled_data_normal.columns)[10:]
columns_correlation_fault= ['Cycle','Phase','Corelation_FaultLocation_FromGenerator',
'Iabc_generator_ContinueComponent_PhaseA','Iabc_generator_ContinueComponent_PhaseB','Iabc_generator_ContinueComponent_PhaseC','Iabc_generator_EffectiveComponent_PhaseA','Iabc_generator_EffectiveComponent_PhaseB','Iabc_generator_EffectiveComponent_PhaseC','Iabc_generator_FirstArmonic_PhaseA_Value','Iabc_generator_FirstArmonic_PhaseA_Angle','Iabc_generator_FirstArmonic_PhaseB_Value','Iabc_generator_FirstArmonic_PhaseB_Angle','Iabc_generator_FirstArmonic_PhaseC_Value','Iabc_generator_FirstArmonic_PhaseC_Angle','Iabc_generator_DeformingComponent_PhaseA','Iabc_generator_DeformingComponent_PhaseB','Iabc_generator_DeformingComponent_PhaseC','Iabc_generator_CoefficientDistorsion_PhaseA','Iabc_generator_CoefficientDistorsion_PhaseB','Iabc_generator_CoefficientDistorsion_PhaseC','Iabc_generator_CoefficientDistorsionContinue_PhaseA','Iabc_generator_CoefficientDistorsionContinue_PhaseB','Iabc_generator_CoefficientDistorsionContinue_PhaseC','Iabc_generator_CoefficientDistorsionFirstArmonic_PhaseA','Iabc_generator_CoefficientDistorsionFirstArmonic_PhaseB','Iabc_generator_CoefficientDistorsionFirstArmonic_PhaseC','Iabc_generator_DirectComponent_Value','Iabc_generator_DirectComponent_Angle','Iabc_generator_InverseComponent_Value','Iabc_generator_InverseComponent_Angle','Iabc_generator_ZeroComponent_Value','Iabc_generator_ZeroComponent_Angle','Iabc_generator_CoefficeintNegativeAsimetry','Iabc_generator_CoefficeintZeroAsimetry',
'Vabc_generator_ContinueComponent_PhaseA','Vabc_generator_ContinueComponent_PhaseB','Vabc_generator_ContinueComponent_PhaseC','Vabc_generator_EffectiveComponent_PhaseA','Vabc_generator_EffectiveComponent_PhaseB','Vabc_generator_EffectiveComponent_PhaseC','Vabc_generator_FirstArmonic_PhaseA_Value','Vabc_generator_FirstArmonic_PhaseA_Angle','Vabc_generator_FirstArmonic_PhaseB_Value','Vabc_generator_FirstArmonic_PhaseB_Angle','Vabc_generator_FirstArmonic_PhaseC_Value','Vabc_generator_FirstArmonic_PhaseC_Angle','Vabc_generator_DeformingComponent_PhaseA','Vabc_generator_DeformingComponent_PhaseB','Vabc_generator_DeformingComponent_PhaseC','Vabc_generator_CoefficientDistorsion_PhaseA','Vabc_generator_CoefficientDistorsion_PhaseB','Vabc_generator_CoefficientDistorsion_PhaseC','Vabc_generator_CoefficientDistorsionContinue_PhaseA','Vabc_generator_CoefficientDistorsionContinue_PhaseB','Vabc_generator_CoefficientDistorsionContinue_PhaseC','Vabc_generator_CoefficientDistorsionFirstArmonic_PhaseA','Vabc_generator_CoefficientDistorsionFirstArmonic_PhaseB','Vabc_generator_CoefficientDistorsionFirstArmonic_PhaseC','Vabc_generator_DirectComponent_Value','Vabc_generator_DirectComponent_Angle','Vabc_generator_InverseComponent_Value','Vabc_generator_InverseComponent_Angle','Vabc_generator_ZeroComponent_Value','Vabc_generator_ZeroComponent_Angle','Vabc_generator_CoefficeintNegativeAsimetry','Vabc_generator_CoefficeintZeroAsimetry',
'Iabc_consumator_ContinueComponent_PhaseA','Iabc_consumator_ContinueComponent_PhaseB','Iabc_consumator_ContinueComponent_PhaseC','Iabc_consumator_EffectiveComponent_PhaseA','Iabc_consumator_EffectiveComponent_PhaseB','Iabc_consumator_EffectiveComponent_PhaseC','Iabc_consumator_FirstArmonic_PhaseA_Value','Iabc_consumator_FirstArmonic_PhaseA_Angle','Iabc_consumator_FirstArmonic_PhaseB_Value','Iabc_consumator_FirstArmonic_PhaseB_Angle','Iabc_consumator_FirstArmonic_PhaseC_Value','Iabc_consumator_FirstArmonic_PhaseC_Angle','Iabc_consumator_DeformingComponent_PhaseA','Iabc_consumator_DeformingComponent_PhaseB','Iabc_consumator_DeformingComponent_PhaseC','Iabc_consumator_CoefficientDistorsion_PhaseA','Iabc_consumator_CoefficientDistorsion_PhaseB','Iabc_consumator_CoefficientDistorsion_PhaseC','Iabc_consumator_CoefficientDistorsionContinue_PhaseA','Iabc_consumator_CoefficientDistorsionContinue_PhaseB','Iabc_consumator_CoefficientDistorsionContinue_PhaseC','Iabc_consumator_CoefficientDistorsionFirstArmonic_PhaseA','Iabc_consumator_CoefficientDistorsionFirstArmonic_PhaseB','Iabc_consumator_CoefficientDistorsionFirstArmonic_PhaseC','Iabc_consumator_DirectComponent_Value','Iabc_consumator_DirectComponent_Angle','Iabc_consumator_InverseComponent_Value','Iabc_consumator_InverseComponent_Angle','Iabc_consumator_ZeroComponent_Value','Iabc_consumator_ZeroComponent_Angle','Iabc_consumator_CoefficeintNegativeAsimetry','Iabc_consumator_CoefficeintZeroAsimetry',
'Vabc_consumator_ContinueComponent_PhaseA','Vabc_consumator_ContinueComponent_PhaseB','Vabc_consumator_ContinueComponent_PhaseC','Vabc_consumator_EffectiveComponent_PhaseA','Vabc_consumator_EffectiveComponent_PhaseB','Vabc_consumator_EffectiveComponent_PhaseC','Vabc_consumator_FirstArmonic_PhaseA_Value','Vabc_consumator_FirstArmonic_PhaseA_Angle','Vabc_consumator_FirstArmonic_PhaseB_Value','Vabc_consumator_FirstArmonic_PhaseB_Angle','Vabc_consumator_FirstArmonic_PhaseC_Value','Vabc_consumator_FirstArmonic_PhaseC_Angle','Vabc_consumator_DeformingComponent_PhaseA','Vabc_consumator_DeformingComponent_PhaseB','Vabc_consumator_DeformingComponent_PhaseC','Vabc_consumator_CoefficientDistorsion_PhaseA','Vabc_consumator_CoefficientDistorsion_PhaseB','Vabc_consumator_CoefficientDistorsion_PhaseC','Vabc_consumator_CoefficientDistorsionContinue_PhaseA','Vabc_consumator_CoefficientDistorsionContinue_PhaseB','Vabc_consumator_CoefficientDistorsionContinue_PhaseC','Vabc_consumator_CoefficientDistorsionFirstArmonic_PhaseA','Vabc_consumator_CoefficientDistorsionFirstArmonic_PhaseB','Vabc_consumator_CoefficientDistorsionFirstArmonic_PhaseC','Vabc_consumator_DirectComponent_Value','Vabc_consumator_DirectComponent_Angle','Vabc_consumator_InverseComponent_Value',
                            'Vabc_consumator_InverseComponent_Angle','Vabc_consumator_ZeroComponent_Value','Vabc_consumator_ZeroComponent_Angle',
                            'Vabc_consumator_CoefficeintNegativeAsimetry','Vabc_consumator_CoefficeintZeroAsimetry']
item_column = ["Cycle","Phase","Corelation_FaultLocation_FromGenerator"]
item_column.extend(columns_to_analyze)
result_corr_scaled_data = list()
result_corr_normal_data = list()
phases_list = list(scaled_data_normal["phase"].unique())

result_corr_scaled_data.append(tuple(item_column))
result_corr_normal_data.append(tuple(item_column))
for cycle in range(0,20):
    for phase in phases_list:
        item_normal = [cycle,phase,1]
        item_scaled = [cycle,phase,1]
        for column in columns_to_analyze:
            aux_scaled = scaled_data_normal[(scaled_data_normal["cycle"] == cycle) & 
                                     (scaled_data_normal["phase"] == phase)][['FaultLocation_FromGenerator',column]].corr()
            aux_normal = data_normal[(data_normal["cycle"] == cycle) & 
                                     (data_normal["phase"] == phase)][['FaultLocation_FromGenerator',column]].corr()
            item_scaled.append(aux_scaled['FaultLocation_FromGenerator'][column])
            item_normal.append(aux_normal['FaultLocation_FromGenerator'][column])

        result_corr_scaled_data.append(tuple(item_scaled))
        result_corr_normal_data.append(tuple(item_normal))
dataFrame_correlation_ScaledData =  pd.DataFrame(data = result_corr_scaled_data[1:], columns = result_corr_scaled_data[0])
dataFrame_correlation_NormalData =  pd.DataFrame(data = result_corr_normal_data[1:], columns = result_corr_normal_data[0])
dataFrame_correlation_ScaledData_fault = dataFrame_correlation_ScaledData[columns_correlation_fault].copy()
dataFrame_correlation_NormalData_fault = dataFrame_correlation_NormalData[columns_correlation_fault].copy()
dataFrame_correlation_NormalData.to_csv('Save_Correlation_Normal_data.csv')
dataFrame_correlation_ScaledData.to_csv('Save_Correlation_Normal_data_SCALED.csv')
dataFrame_correlation_ScaledData_fault.to_csv('Save_Correlation_Fault_data.csv')
dataFrame_correlation_NormalData_fault.to_csv('Save_Correlation_Fault_data_SCALED.csv')

#pd.DataFrame(data = result_corr_scaled_data[1:], columns = result_corr_scaled_data[0])

#plt.figure(figsize=(8, 6))
#plt.plot(aux["Cycle"],aux[column], linewidth=3)

#plt.title('Case: '+ phase+ '\n'+ column)
#plt.xlabel('Number of Cycles')
#plt.ylabel('Correlation Factor')
#plt.grid(True,color='#000000')
#plt.savefig(outputImage)
#plt.figure().clear() 
#plt.cla()
#plt.clf()
#plt.close()

fig=plt.figure()
axes=fig.add_axes([0,0,1,1])
axes.plot(aux["Cycle"],aux["Iabc_generator_ContinueComponent_PhaseA"], linewidth=3)
axes.set_title('Case: '+ 'A'+ '\n'+'Iabc_generator_ContinueComponent_PhaseA')
axes.set_xlabel('Number of Cycles')
axes.set_ylabel('Correlation Factor')
axes.grid(True,color='#000000')
plt.close()

def SaveLineGraphsCorrelation(data_frame, outputDir="Output"):
    isExist = os.path.exists(outputDir)
    if not isExist:
        os.makedirs(outputDir)    
    dir_path = os.path.join(os.getcwd(),outputDir)
    phases_list = list(data_frame["Phase"].unique())
    for phase in phases_list:
        aux = data_frame[(data_frame["Phase"] == phase)]
        isExist = os.path.exists(os.path.join(os.path.join(os.getcwd(),outputDir)
                                                      ,phase))
        if not isExist:
            os.makedirs(os.path.join(os.path.join(os.getcwd(),outputDir)
                                                      ,phase))
        for column in list(aux.columns)[3:]:
            fileName = 'Case_'+phase+'_'+column + ".png"
            outputImage = os.path.join(os.path.join(os.path.join(os.getcwd(),outputDir),str(phase)), fileName)
            plt.clf()
            plt.figure()
            axes=plt.axes([0,0,1,1])
            axes.plot(aux["Cycle"],aux[str(column)], linewidth=3)
            axes.set_title('Case: '+ str(phase) + '\n'+ str(column) )
            axes.set_xlabel('Number of Cycles')
            axes.set_ylabel('Correlation Factor')
            axes.grid(True,color='#000000')
            plt.savefig(outputImage)
            plt.close()


SaveLineGraphsCorrelation(dataFrame_correlation_ScaledData, outputDir="dataFrame_correlation_ScaledData")

SaveLineGraphsCorrelation(dataFrame_correlation_NormalData, outputDir="dataFrame_correlation_NormalData")
SaveLineGraphsCorrelation(dataFrame_correlation_ScaledData_fault, outputDir="dataFrame_correlation_ScaledData_fault")
SaveLineGraphsCorrelation(dataFrame_correlation_NormalData_fault, outputDir="dataFrame_correlation_NormalData_fault")
