==> 01_Compute and display date/03_displaySeabornGraphs_part2.ipynb <==
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
%matplotlib inline
import seaborn as sns
import os

data_normal = pd.read_csv('Save_Normal_data.csv')
scaled_data_normal = pd.read_csv('Save_Normal_data_SCALED.csv')
data_normal_fault = pd.read_csv('Save_Fault_data.csv')
scaled_data_normal_fault = pd.read_csv('Save_Fault_data_SCALED.csv')

dataFrame_correlation_ScaledData = pd.read_csv('Save_Correlation_Normal_data_SCALED.csv')
dataFrame_correlation_NormalData = pd.read_csv('Save_Correlation_Normal_data.csv')
dataFrame_correlation_ScaledData_fault = pd.read_csv('Save_Correlation_Fault_data_SCALED.csv')
dataFrame_correlation_NormalData_fault = pd.read_csv('Save_Correlation_Fault_data.csv')

aux = scaled_data_normal[scaled_data_normal["cycle"]==19]

sns.histplot(data = aux, x = 'FaultLocation_FromGenerator'
             , y='Iabc_generator_ContinueComponent_PhaseA'
            ,hue='Type')

sns.histplot(data = aux, x = 'FaultLocation_FromGenerator'
             , y='Iabc_generator_ContinueComponent_PhaseA'
            ,hue='phase')

sns.jointplot(x = 'FaultLocation_FromGenerator', y='Iabc_generator_ContinueComponent_PhaseA', data = aux, kind='scatter')

sns.jointplot(x = 'FaultLocation_FromGenerator', y='Iabc_generator_ContinueComponent_PhaseA'
              , data = aux, kind='reg')

sns.scatterplot(x = 'FaultLocation_FromGenerator', y='Iabc_generator_ContinueComponent_PhaseA'
              , data = aux, hue='phase')

sns.scatterplot(x = 'FaultLocation_FromGenerator', y='Iabc_generator_ContinueComponent_PhaseA'
              , data = aux, hue='Type')

sns.scatterplot(x = 'FaultLocation_FromGenerator', y='Iabc_generator_ContinueComponent_PhaseA'
              , data = aux[aux["Type"]!='HighVoltage_LongLine300'], hue='Type')

sns.heatmap(aux[list(aux.columns)[11:]].corr(), annot=False)

copy = scaled_data_normal.copy()
copy['FaultLocation_FromGenerator'] = copy['FaultLocation_FromGenerator'].apply(lambda x: np.round(x,2))

pt1 = copy[copy['phase']=='A'].pivot_table(index='FaultLocation_FromGenerator', columns='cycle',
                                     values='Iabc_generator_ContinueComponent_PhaseA' )

sns.heatmap(pt1,cmap='coolwarm', fmt='.2f')

pt2 = copy[copy["cycle"]==19].pivot_table(index='phase', columns='FaultLocation_FromGenerator',
                                     values='Iabc_generator_ContinueComponent_PhaseA' )

sns.heatmap(pt2,cmap='coolwarm', fmt='.4f')

sns.kdeplot(x=aux['FaultLocation_FromGenerator'], y=aux['Iabc_generator_ContinueComponent_PhaseA'])

g = sns.FacetGrid(data=aux,col='phase',row='Type')
g.map(sns.scatterplot,'FaultLocation_FromGenerator','Iabc_generator_ContinueComponent_PhaseA')


g1 = sns.FacetGrid(data=aux,col='phase')
g1.map(sns.scatterplot,'FaultLocation_FromGenerator','Iabc_generator_ContinueComponent_PhaseA')

g1 = sns.FacetGrid(data=aux,col='Type')
g1.map(sns.scatterplot,'FaultLocation_FromGenerator','Iabc_generator_ContinueComponent_PhaseA')

from contextlib import closing

def SaveHistPlot(data_frame, cycle, outputDir="Output", chosen_hue='Type', start = 0, end = 20):
    isExist = os.path.exists(outputDir)
    if not isExist:
        os.makedirs(outputDir)    
    dir_path = os.path.join(os.getcwd(),outputDir)
    name_c = "Cycle "+ str(cycle)
    isExist = os.path.exists("./"+outputDir+"/"+name_c)
    if not isExist:
        os.makedirs("./"+outputDir+"/"+name_c) 
        
    aux = data_frame[data_frame["cycle"]==cycle]
    list_col = list(aux.columns)[11+start:11+end]
    for column in list_col:
        seaborn_elem = sns.histplot(data = aux[['FaultLocation_FromGenerator',column,chosen_hue]], 
                                    x = 'FaultLocation_FromGenerator', y=column,hue=chosen_hue)
        fig = seaborn_elem.get_figure()
        dir_name = os.path.join(os.path.join(os.path.join(os.getcwd(),outputDir),name_c),str(column) + ".png")
        fig.savefig(dir_name)
        plt.clf()
        plt.close()

def SaveJoinPlot(data_frame, cycle, outputDir="Output", chosen_kind='scatter', start = 0, end = 20):
    name_c = "Cycle "+ str(cycle)
    isExist = os.path.exists("./"+outputDir+"/"+name_c)
    if not isExist:
        os.makedirs("./"+outputDir+"/"+name_c) 
    aux = data_frame[data_frame["cycle"]==cycle]
    list_col = list(aux.columns)[11+start:11+end]
    for column in list_col:
        seaborn_elem = sns.jointplot(x = 'FaultLocation_FromGenerator', y=column, 
                                     data = aux[['FaultLocation_FromGenerator',column]],kind = chosen_kind)
        fig = seaborn_elem.get_figure()
        dir_name = os.path.join(os.path.join(os.path.join(os.getcwd(),outputDir),name_c),str(column) + ".png")
        fig.savefig(dir_name)
        plt.clf()
        plt.close()
        
def SaveScatterPlot(data_frame, cycle, outputDir="Output", chosen_hue='Type', start = 0, end = 20):
    isExist = os.path.exists(outputDir)
    if not isExist:
        os.makedirs(outputDir)    
    dir_path = os.path.join(os.getcwd(),outputDir)
    name_c = "Cycle "+ str(cycle)
    isExist = os.path.exists("./"+outputDir+"/"+name_c)
    if not isExist:
        os.makedirs("./"+outputDir+"/"+name_c) 
        
    aux = data_frame[data_frame["cycle"]==cycle]
    list_col = list(aux.columns)[11+start:11+end]
    for column in list_col:
        seaborn_elem = sns.scatterplot(data = aux[['FaultLocation_FromGenerator',column,chosen_hue]], 
                                    x = 'FaultLocation_FromGenerator', y=column,hue=chosen_hue)
        fig = seaborn_elem.get_figure()
        dir_name = os.path.join(os.path.join(os.path.join(os.getcwd(),outputDir),name_c),str(column) + ".png")
        fig.savefig(dir_name)
        plt.clf()
        plt.close()
        
def SaveKdePlot(data_frame, cycle, outputDir="Output", start = 0, end = 20):
    isExist = os.path.exists(outputDir)
    if not isExist:
        os.makedirs(outputDir)    
    dir_path = os.path.join(os.getcwd(),outputDir)
    name_c = "Cycle "+ str(cycle)
    isExist = os.path.exists("./"+outputDir+"/"+name_c)
    if not isExist:
        os.makedirs("./"+outputDir+"/"+name_c)
    aux = data_frame[data_frame["cycle"]==cycle]
    list_col = list(aux.columns)[11+start:11+end]    
    for column in list_col:
        seaborn_elem = sns.kdeplot(x=aux['FaultLocation_FromGenerator'], y=aux[column])
        fig = seaborn_elem.get_figure()
        dir_name = os.path.join(os.path.join(os.path.join(os.getcwd(),outputDir),name_c),str(column) + ".png")
        fig.savefig(dir_name)
        plt.clf()
        plt.close()
        
def SaveFacetGridPlot(data_frame, cycle, outputDir="Output", chosen_hue='Type', start = 0, end = 20):
    isExist = os.path.exists(outputDir)
    if not isExist:
        os.makedirs(outputDir)    
    dir_path = os.path.join(os.getcwd(),outputDir)
    name_c = "Cycle "+ str(cycle)
    isExist = os.path.exists("./"+outputDir+"/"+name_c)
    if not isExist:
        os.makedirs("./"+outputDir+"/"+name_c)
    aux = data_frame[data_frame["cycle"]==cycle]
    list_col = list(aux.columns)[11+start:11+end] 
    for column in list_col:
        seaborn_elem = sns.FacetGrid(data=aux[[chosen_hue, column, 'FaultLocation_FromGenerator']],col=chosen_hue) 
        seaborn_elem.map(sns.scatterplot,'FaultLocation_FromGenerator',column)
        dir_name = os.path.join(os.path.join(os.path.join(os.getcwd(),outputDir),name_c),str(column) + ".png")
        seaborn_elem.savefig(dir_name)
        plt.clf()
        plt.close()  

def SaveHeatMap_v1(data_frame, outputDir="Output"):
    isExist = os.path.exists(outputDir)
    if not isExist:
        os.makedirs(outputDir)    
    dir_path = os.path.join(os.getcwd(),outputDir)
    for cycle in range(0,20):
        aux = data_frame[data_frame["cycle"]==cycle]
        seaborn_elem = sns.heatmap(aux[list(aux.columns)[11:]].corr(), annot=False)
        fig = seaborn_elem.get_figure()
        dir_name = os.path.join(os.path.join(os.path.join(os.getcwd(),outputDir),name_c),"Cycle"+str(cycle) + ".png")
        fig.savefig(dir_name)
        plt.clf()
        plt.close() 

def SaveHeatMap_v2_phase(data_frame, outputDir="Output", ch_phase='A', start = 0, end = 20):
    isExist = os.path.exists(outputDir)
    if not isExist:
        os.makedirs(outputDir)    
    dir_path = os.path.join(os.getcwd(),outputDir)
    aux = data_frame[aux["phase"]==ch_phase]
    aux['FaultLocation_FromGenerator'] = aux['FaultLocation_FromGenerator'].apply(lambda x: np.round(x,2))
    list_col = list(aux.columns)[11+start:11+end] 
    for column in list_col:
        pt1 = copy.pivot_table(index='FaultLocation_FromGenerator', columns='cycle',
                                     values=column )
        seaborn_elem = sns.heatmap(pt1,cmap='coolwarm', fmt='.2f')
        fig = seaborn_elem.get_figure()
        dir_name = os.path.join(os.path.join(os.path.join(os.getcwd(),outputDir),name_c),str(column) + ".png")
        fig.savefig(dir_name)
        plt.clf()
        plt.close()

def SaveHeatMap_v2_cycle(data_frame, outputDir="Output", start = 0, end = 20, ch_cycle = 19):
    isExist = os.path.exists(outputDir)
    if not isExist:
        os.makedirs(outputDir)    
    dir_path = os.path.join(os.getcwd(),outputDir)
    aux = data_frame[aux["cycle"]==19]
    aux['FaultLocation_FromGenerator'] = aux['FaultLocation_FromGenerator'].apply(lambda x: np.round(x,2))    
    list_col = list(aux.columns)[11+start:11+end] 
    for column in list_col:
        pt1 = copy.pivot_table(index='FaultLocation_FromGenerator', columns='phase',
                                     values=column )
        seaborn_elem = sns.heatmap(pt1,cmap='coolwarm', fmt='.2f')
        fig = seaborn_elem.get_figure()
        dir_name = os.path.join(os.path.join(os.path.join(os.getcwd(),outputDir),name_c),str(column) + ".png")
        fig.savefig(dir_name)
        plt.clf()
        plt.close()   

copy = scaled_data_normal.copy()
copy['FaultLocation_FromGenerator'] = copy['FaultLocation_FromGenerator'].apply(lambda x: np.round(x,2))
pt1 = copy.pivot_table(index='FaultLocation_FromGenerator', columns='cycle',
                                     values='Iabc_generator_ContinueComponent_PhaseA' )
sns.heatmap(pt1,cmap='coolwarm', fmt='.2f')

pt2 = copy[copy["cycle"]==19].pivot_table(index='FaultLocation_FromGenerator', columns='phase',
                                     values='Iabc_generator_ContinueComponent_PhaseA' )
sns.heatmap(pt2,cmap='coolwarm', fmt='.4f')



pt1 = copy[copy['phase']=='A'].pivot_table(index='FaultLocation_FromGenerator', columns='cycle',
                                     values='Iabc_generator_ContinueComponent_PhaseA' )

sns.heatmap(pt1,cmap='coolwarm', fmt='.2f')
plt.close()

def SingleSaveHistPlot(data_frame, column, cycle, outputDir="Output", chosen_hue='Type'):
    isExist = os.path.exists(outputDir)
    if not isExist:
        os.makedirs(outputDir)    
    dir_path = os.path.join(os.getcwd(),outputDir)
    name_c = "Cycle "+ str(cycle)
    isExist = os.path.exists("./"+outputDir+"/"+name_c)
    if not isExist:
        os.makedirs("./"+outputDir+"/"+name_c) 
    #aux = data_frame
    aux = data_frame[data_frame["cycle"]==cycle]
    seaborn_elem = sns.histplot(data = aux[['FaultLocation_FromGenerator',column,chosen_hue]], 
                                    x = 'FaultLocation_FromGenerator', y=column,hue=chosen_hue)
    #seaborn_elem.fig.suptitle("HistPlot - Cycle "+str(cycle))
    #seaborn_elem.set_title("HistPlot - Cycle "+str(cycle))
    fig = seaborn_elem.get_figure()
    dir_name = os.path.join(os.path.join(os.path.join(os.getcwd(),outputDir),name_c),str(column) + ".png")
    #fig.set_title("HistPlot - Cycle "+str(cycle))
    fig.savefig(dir_name, bbox_inches="tight")
    plt.clf()
    plt.close()

def SingleSaveHistPlot_bins(data_frame, column, cycle, outputDir="Output", chosen_hue='Type', _bins=30):
    isExist = os.path.exists(outputDir)
    if not isExist:
        os.makedirs(outputDir)    
    dir_path = os.path.join(os.getcwd(),outputDir)
    name_c = "Cycle "+ str(cycle)
    isExist = os.path.exists("./"+outputDir+"/"+name_c)
    if not isExist:
        os.makedirs("./"+outputDir+"/"+name_c) 
    #aux = data_frame
    aux = data_frame[data_frame["cycle"]==cycle]
    seaborn_elem = sns.histplot(data = aux[['FaultLocation_FromGenerator',column,chosen_hue]], 
                                    x = 'FaultLocation_FromGenerator', y=column,hue=chosen_hue, bins=_bins)
    #seaborn_elem.fig.suptitle("HistPlot - Cycle "+str(cycle))
    #seaborn_elem.set_title("HistPlot - Cycle "+str(cycle))
    fig = seaborn_elem.get_figure()
    dir_name = os.path.join(os.path.join(os.path.join(os.getcwd(),outputDir),name_c),str(column) + ".png")
    fig.savefig(dir_name, bbox_inches="tight")
    plt.clf()
    plt.close()

def SingleSaveJoinPlot(data_frame, column, cycle, outputDir="Output", chosen_kind='scatter'):
    name_c = "Cycle "+ str(cycle)
    isExist = os.path.exists("./"+outputDir+"/"+name_c)
    if not isExist:
        os.makedirs("./"+outputDir+"/"+name_c) 
    #aux = data_frame
    aux = data_frame[data_frame["cycle"]==cycle]
    seaborn_elem = sns.jointplot(x = 'FaultLocation_FromGenerator', y=column, 
                                     data = aux[['FaultLocation_FromGenerator',column]],kind = chosen_kind)
    seaborn_elem.fig.suptitle("JoinPlot - Cycle "+str(cycle))
    dir_name = os.path.join(os.path.join(os.path.join(os.getcwd(),outputDir),name_c),str(column) + ".png")
    
    seaborn_elem.fig.savefig(dir_name, bbox_inches="tight")
    
    plt.clf()
    plt.close()
    
def SingleSaveJoinPlot_bins(data_frame, column, cycle, outputDir="Output", chosen_kind='scatter', bins=30):
    name_c = "Cycle "+ str(cycle)
    isExist = os.path.exists("./"+outputDir+"/"+name_c)
    if not isExist:
        os.makedirs("./"+outputDir+"/"+name_c) 
    #aux = data_frame
    aux = data_frame[data_frame["cycle"]==cycle]
    seaborn_elem = sns.jointplot(x = 'FaultLocation_FromGenerator', y=column, 
                                     data = aux[['FaultLocation_FromGenerator',column]],kind = chosen_kind, s=bins)
    seaborn_elem.fig.suptitle("JoinPlot - Cycle "+str(cycle))
    dir_name = os.path.join(os.path.join(os.path.join(os.getcwd(),outputDir),name_c),str(column) + ".png")
    
    seaborn_elem.fig.savefig(dir_name, bbox_inches="tight")
    
    plt.clf()
    plt.close()
    
def SingleSaveScatterPlot(data_frame, column, cycle, outputDir="Output", chosen_hue='Type'):
    isExist = os.path.exists(outputDir)
    if not isExist:
        os.makedirs(outputDir)    
    dir_path = os.path.join(os.getcwd(),outputDir)
    name_c = "Cycle "+ str(cycle)
    isExist = os.path.exists("./"+outputDir+"/"+name_c)
    if not isExist:
        os.makedirs("./"+outputDir+"/"+name_c) 
        
    #aux = data_frame
    aux = data_frame[data_frame["cycle"]==cycle]
    seaborn_elem = sns.scatterplot(data = aux[['FaultLocation_FromGenerator',column,chosen_hue]], 
                                    x = 'FaultLocation_FromGenerator', y=column,hue=chosen_hue)
    seaborn_elem.set_title("ScatterPlot - Cycle "+str(cycle))
    fig = seaborn_elem.get_figure()
    dir_name = os.path.join(os.path.join(os.path.join(os.getcwd(),outputDir),name_c),str(column) + ".png")
    fig.savefig(dir_name, bbox_inches="tight")
    plt.clf()
    plt.close()

def SingleSaveKdePlot(data_frame, column, cycle, outputDir="Output"):
    isExist = os.path.exists(outputDir)
    if not isExist:
        os.makedirs(outputDir)    
    dir_path = os.path.join(os.getcwd(),outputDir)
    name_c = "Cycle "+ str(cycle)
    isExist = os.path.exists("./"+outputDir+"/"+name_c)
    if not isExist:
        os.makedirs("./"+outputDir+"/"+name_c)
    #aux = data_frame
    aux = data_frame[data_frame["cycle"]==cycle]
    seaborn_elem = sns.kdeplot(x=aux['FaultLocation_FromGenerator'], y=aux[column])
    seaborn_elem.set_title("KdePlot - Cycle "+str(cycle))
    fig = seaborn_elem.get_figure()
    dir_name = os.path.join(os.path.join(os.path.join(os.getcwd(),outputDir),name_c),str(column) + ".png")
    fig.savefig(dir_name, bbox_inches="tight")
    plt.clf()
    plt.close()

def SingleSaveFacetGridPlot(data_frame, column, cycle, outputDir="Output", chosen_hue='Type'):
    isExist = os.path.exists(outputDir)
    if not isExist:
        os.makedirs(outputDir)    
    dir_path = os.path.join(os.getcwd(),outputDir)
    name_c = "Cycle "+ str(cycle)
    isExist = os.path.exists("./"+outputDir+"/"+name_c)
    if not isExist:
        os.makedirs("./"+outputDir+"/"+name_c)
    #aux = data_frame
    aux = data_frame[data_frame["cycle"]==cycle]
    seaborn_elem = sns.FacetGrid(data=aux[[chosen_hue, column, 'FaultLocation_FromGenerator']],col=chosen_hue) 
    seaborn_elem.fig.suptitle("FacetGrid - Cycle "+str(cycle))
    seaborn_elem.map(sns.scatterplot,'FaultLocation_FromGenerator',column)
    dir_name = os.path.join(os.path.join(os.path.join(os.getcwd(),outputDir),name_c),str(column) + ".png")
    seaborn_elem.fig.savefig(dir_name, bbox_inches="tight")
    plt.clf()
    plt.close() 

def SingleSaveHeatMap_v1(data_frame, cycle, outputDir="Output"):
    isExist = os.path.exists(outputDir)
    if not isExist:
        os.makedirs(outputDir)    
    dir_path = os.path.join(os.getcwd(),outputDir)
    #aux = data_frame
    aux = data_frame[data_frame["cycle"]==cycle]
    
    seaborn_elem = sns.heatmap(aux[list(aux.columns)[11:]].corr(), annot=False)
    seaborn_elem.set_title("HeatMap - Cycle "+str(cycle))
    fig = seaborn_elem.get_figure()
    plt.figure(figsize = (12,12))
    dir_name = os.path.join(os.path.join(os.getcwd(),outputDir),"Cycle"+str(cycle) + ".png")
    fig.savefig(dir_name, bbox_inches="tight")
    plt.clf()
    plt.close() 
    
def SingleSaveHeatMap_v1_output(data_frame, cycle, outputDir="Output"):
    isExist = os.path.exists(outputDir)
    if not isExist:
        os.makedirs(outputDir)    
    dir_path = os.path.join(os.getcwd(),outputDir)
    #aux = data_frame
    aux = data_frame[data_frame["cycle"]==cycle]
    
    seaborn_elem = sns.heatmap(aux[list(aux.columns)[11:]].corr(), annot=False)
    seaborn_elem.set_title("HeatMap - Cycle "+str(cycle))

def SingleSaveHeatMap_v2_phase(data_frame, column, outputDir="Output", ch_phase='A', pt_column='cycle'):
    isExist = os.path.exists(outputDir)
    if not isExist:
        os.makedirs(outputDir)    
    dir_path = os.path.join(os.getcwd(),outputDir)
    aux = data_frame[data_frame["phase"]==ch_phase]
    copy = aux.copy()
    copy['FaultLocation_FromGenerator'] = aux['FaultLocation_FromGenerator'].apply(lambda x: np.round(x,2))
    
    pt1 = copy.pivot_table(index='FaultLocation_FromGenerator', columns=pt_column,
                                     values=column )
    seaborn_elem = sns.heatmap(pt1,cmap='coolwarm', fmt='.2f')
    seaborn_elem.set_title("HeatMap - Phase "+str(ch_phase) + ' - ' + str(column))
    fig = seaborn_elem.get_figure()
    name_c = "Phase"+ch_phase
    isExist = os.path.exists("./"+outputDir+"/"+name_c)
    if not isExist:
        os.makedirs("./"+outputDir+"/"+name_c)
    dir_name = os.path.join(os.path.join(os.path.join(os.getcwd(),outputDir),name_c),str(column) + ".png")
    fig.savefig(dir_name, bbox_inches="tight")
    plt.clf()
    plt.close()
    
def SingleSaveHeatMap_v2_cycle(data_frame, column, outputDir="Output", ch_cycle = 19, pt_column='phase'):
    isExist = os.path.exists(outputDir)
    if not isExist:
        os.makedirs(outputDir)    
    dir_path = os.path.join(os.getcwd(),outputDir)
    aux = data_frame[data_frame["cycle"]==ch_cycle]
    copy = aux.copy()
    copy['FaultLocation_FromGenerator'] = aux['FaultLocation_FromGenerator'].apply(lambda x: np.round(x,2))

    pt1 = copy.pivot_table(index='FaultLocation_FromGenerator', columns=pt_column,
                                     values=column )
    seaborn_elem = sns.heatmap(pt1,cmap='coolwarm', fmt='.2f')
    seaborn_elem.set_title("HeatMap - Cycle "+str(ch_cycle) + ' - ' + str(column))
    name_c = "Cycle"+str(ch_cycle)
    isExist = os.path.exists("./"+outputDir+"/"+name_c)
    if not isExist:
        os.makedirs("./"+outputDir+"/"+name_c)    
    fig = seaborn_elem.get_figure()
    dir_name = os.path.join(os.path.join(os.path.join(os.getcwd(),outputDir),name_c),str(column) + ".png")
    fig.savefig(dir_name, bbox_inches="tight")
    plt.clf()
    plt.close()  

def SingleSaveHeatMap_v2_type(data_frame, column, outputDir="Output", ch_type = '', pt_column='cycle'):
    isExist = os.path.exists(outputDir)
    if not isExist:
        os.makedirs(outputDir)    
    dir_path = os.path.join(os.getcwd(),outputDir)
    aux = data_frame[data_frame["Type"]==ch_type]
    copy = aux.copy()
    copy['FaultLocation_FromGenerator'] = aux['FaultLocation_FromGenerator'].apply(lambda x: np.round(x,2)) 

    pt1 = copy.pivot_table(index='FaultLocation_FromGenerator', columns=pt_column,
                                     values=column )
    seaborn_elem = sns.heatmap(pt1,cmap='coolwarm', fmt='.2f')
    seaborn_elem.set_title("HeatMap - Type "+str(ch_type) + ' - ' + str(column))
    fig = seaborn_elem.get_figure()
    name_c = "Type"+ch_type
    
    isExist = os.path.exists("./"+outputDir+"/"+name_c)
    if not isExist:
        os.makedirs("./"+outputDir+"/"+name_c)

    dir_name = os.path.join(os.path.join(os.path.join(os.getcwd(),outputDir),name_c),str(column) + ".png")
    fig.savefig(dir_name, bbox_inches="tight")
    plt.clf()
    plt.close()  

data_normal["phase"].unique()

data_normal["Type"].unique()

#_output_dir="data_normal_Scatter"

_output_dir="data_normal_Type"
_hue='Type'

#_output_dir="data_normal_Phase"
#_hue='phase'

#_output_dir="data_normal_Cycle"
#_hue='cycle'

_columns = list(data_normal.columns)[11:]
_chosen_kind = 'scatter'
_cycle=19
c_bins=30
_ch_phase = 'C'
_ch_type = 'MediumVoltage_ShortLine'


#   SingleSaveHistPlot(data_normal, '', _cycle, outputDir=_output_dir, chosen_hue=_hue)
#   SingleSaveHistPlot_bins(data_normal, '', _cycle, outputDir=_output_dir, chosen_hue=_hue, _bins=c_bins)


_output_dir="data_normal_Phase"
_hue='phase'

_cycle=19
SingleSaveHistPlot_bins(data_normal, 'Iabc_generator_ContinueComponent_PhaseA', _cycle, outputDir=_output_dir, chosen_hue=_hue, _bins=c_bins)
SingleSaveHistPlot_bins(data_normal, 'Iabc_generator_ContinueComponent_PhaseB', _cycle, outputDir=_output_dir, chosen_hue=_hue, _bins=c_bins)
SingleSaveHistPlot_bins(data_normal, 'Iabc_generator_ContinueComponent_PhaseC', _cycle, outputDir=_output_dir, chosen_hue=_hue, _bins=c_bins)
SingleSaveHistPlot_bins(data_normal, 'Iabc_generator_EffectiveComponent_PhaseA', _cycle, outputDir=_output_dir, chosen_hue=_hue, _bins=c_bins)
SingleSaveHistPlot_bins(data_normal, 'Iabc_generator_EffectiveComponent_PhaseB', _cycle, outputDir=_output_dir, chosen_hue=_hue, _bins=c_bins)
SingleSaveHistPlot_bins(data_normal, 'Iabc_generator_EffectiveComponent_PhaseC', _cycle, outputDir=_output_dir, chosen_hue=_hue, _bins=c_bins)
SingleSaveHistPlot_bins(data_normal, 'Iabc_generator_FirstArmonic_PhaseA_Value', _cycle, outputDir=_output_dir, chosen_hue=_hue, _bins=c_bins)
SingleSaveHistPlot_bins(data_normal, 'Iabc_generator_FirstArmonic_PhaseA_Angle', _cycle, outputDir=_output_dir, chosen_hue=_hue, _bins=c_bins)
SingleSaveHistPlot_bins(data_normal, 'Iabc_generator_FirstArmonic_PhaseB_Value', _cycle, outputDir=_output_dir, chosen_hue=_hue, _bins=c_bins)
SingleSaveHistPlot_bins(data_normal, 'Iabc_generator_FirstArmonic_PhaseB_Angle', _cycle, outputDir=_output_dir, chosen_hue=_hue, _bins=c_bins)
SingleSaveHistPlot_bins(data_normal, 'Iabc_generator_FirstArmonic_PhaseC_Value', _cycle, outputDir=_output_dir, chosen_hue=_hue, _bins=c_bins)
SingleSaveHistPlot_bins(data_normal, 'Iabc_generator_FirstArmonic_PhaseC_Angle', _cycle, outputDir=_output_dir, chosen_hue=_hue, _bins=c_bins)
SingleSaveHistPlot_bins(data_normal, 'Iabc_generator_DeformingComponent_PhaseA', _cycle, outputDir=_output_dir, chosen_hue=_hue, _bins=c_bins)
SingleSaveHistPlot_bins(data_normal, 'Iabc_generator_DeformingComponent_PhaseB', _cycle, outputDir=_output_dir, chosen_hue=_hue, _bins=c_bins)
SingleSaveHistPlot_bins(data_normal, 'Iabc_generator_DeformingComponent_PhaseC', _cycle, outputDir=_output_dir, chosen_hue=_hue, _bins=c_bins)
SingleSaveHistPlot_bins(data_normal, 'Iabc_generator_CoefficientDistorsion_PhaseA', _cycle, outputDir=_output_dir, chosen_hue=_hue, _bins=c_bins)
SingleSaveHistPlot_bins(data_normal, 'Iabc_generator_CoefficientDistorsion_PhaseB', _cycle, outputDir=_output_dir, chosen_hue=_hue, _bins=c_bins)
SingleSaveHistPlot_bins(data_normal, 'Iabc_generator_CoefficientDistorsion_PhaseC', _cycle, outputDir=_output_dir, chosen_hue=_hue, _bins=c_bins)
SingleSaveHistPlot_bins(data_normal, 'Iabc_generator_CoefficientDistorsionContinue_PhaseA', _cycle, outputDir=_output_dir, chosen_hue=_hue, _bins=c_bins)
SingleSaveHistPlot_bins(data_normal, 'Iabc_generator_CoefficientDistorsionContinue_PhaseB', _cycle, outputDir=_output_dir, chosen_hue=_hue, _bins=c_bins)
SingleSaveHistPlot_bins(data_normal, 'Iabc_generator_CoefficientDistorsionContinue_PhaseC', _cycle, outputDir=_output_dir, chosen_hue=_hue, _bins=c_bins)
SingleSaveHistPlot_bins(data_normal, 'Iabc_generator_DirectComponent_Value', _cycle, outputDir=_output_dir, chosen_hue=_hue, _bins=c_bins)
SingleSaveHistPlot_bins(data_normal, 'Iabc_generator_DirectComponent_Angle', _cycle, outputDir=_output_dir, chosen_hue=_hue, _bins=c_bins)
SingleSaveHistPlot_bins(data_normal, 'Iabc_generator_InverseComponent_Value', _cycle, outputDir=_output_dir, chosen_hue=_hue, _bins=c_bins)
SingleSaveHistPlot_bins(data_normal, 'Iabc_generator_InverseComponent_Angle', _cycle, outputDir=_output_dir, chosen_hue=_hue, _bins=c_bins)
SingleSaveHistPlot_bins(data_normal, 'Iabc_generator_ZeroComponent_Value', _cycle, outputDir=_output_dir, chosen_hue=_hue, _bins=c_bins)
SingleSaveHistPlot_bins(data_normal, 'Iabc_generator_ZeroComponent_Angle', _cycle, outputDir=_output_dir, chosen_hue=_hue, _bins=c_bins)
SingleSaveHistPlot_bins(data_normal, 'Vabc_generator_ContinueComponent_PhaseA', _cycle, outputDir=_output_dir, chosen_hue=_hue, _bins=c_bins)
SingleSaveHistPlot_bins(data_normal, 'Vabc_generator_ContinueComponent_PhaseB', _cycle, outputDir=_output_dir, chosen_hue=_hue, _bins=c_bins)
SingleSaveHistPlot_bins(data_normal, 'Vabc_generator_ContinueComponent_PhaseC', _cycle, outputDir=_output_dir, chosen_hue=_hue, _bins=c_bins)
SingleSaveHistPlot_bins(data_normal, 'Vabc_generator_EffectiveComponent_PhaseA', _cycle, outputDir=_output_dir, chosen_hue=_hue, _bins=c_bins)
SingleSaveHistPlot_bins(data_normal, 'Vabc_generator_EffectiveComponent_PhaseB', _cycle, outputDir=_output_dir, chosen_hue=_hue, _bins=c_bins)
SingleSaveHistPlot_bins(data_normal, 'Vabc_generator_EffectiveComponent_PhaseC', _cycle, outputDir=_output_dir, chosen_hue=_hue, _bins=c_bins)
SingleSaveHistPlot_bins(data_normal, 'Vabc_generator_FirstArmonic_PhaseA_Value', _cycle, outputDir=_output_dir, chosen_hue=_hue, _bins=c_bins)
SingleSaveHistPlot_bins(data_normal, 'Vabc_generator_FirstArmonic_PhaseA_Angle', _cycle, outputDir=_output_dir, chosen_hue=_hue, _bins=c_bins)
SingleSaveHistPlot_bins(data_normal, 'Vabc_generator_FirstArmonic_PhaseB_Value', _cycle, outputDir=_output_dir, chosen_hue=_hue, _bins=c_bins)
SingleSaveHistPlot_bins(data_normal, 'Vabc_generator_FirstArmonic_PhaseB_Angle', _cycle, outputDir=_output_dir, chosen_hue=_hue, _bins=c_bins)
SingleSaveHistPlot_bins(data_normal, 'Vabc_generator_FirstArmonic_PhaseC_Value', _cycle, outputDir=_output_dir, chosen_hue=_hue, _bins=c_bins)
SingleSaveHistPlot_bins(data_normal, 'Vabc_generator_FirstArmonic_PhaseC_Angle', _cycle, outputDir=_output_dir, chosen_hue=_hue, _bins=c_bins)
SingleSaveHistPlot_bins(data_normal, 'Vabc_generator_DeformingComponent_PhaseA', _cycle, outputDir=_output_dir, chosen_hue=_hue, _bins=c_bins)
SingleSaveHistPlot_bins(data_normal, 'Vabc_generator_DeformingComponent_PhaseB', _cycle, outputDir=_output_dir, chosen_hue=_hue, _bins=c_bins)
SingleSaveHistPlot_bins(data_normal, 'Vabc_generator_DeformingComponent_PhaseC', _cycle, outputDir=_output_dir, chosen_hue=_hue, _bins=c_bins)
SingleSaveHistPlot_bins(data_normal, 'Vabc_generator_CoefficientDistorsion_PhaseA', _cycle, outputDir=_output_dir, chosen_hue=_hue, _bins=c_bins)
SingleSaveHistPlot_bins(data_normal, 'Vabc_generator_CoefficientDistorsion_PhaseB', _cycle, outputDir=_output_dir, chosen_hue=_hue, _bins=c_bins)
SingleSaveHistPlot_bins(data_normal, 'Vabc_generator_CoefficientDistorsion_PhaseC', _cycle, outputDir=_output_dir, chosen_hue=_hue, _bins=c_bins)
SingleSaveHistPlot_bins(data_normal, 'Vabc_generator_CoefficientDistorsionContinue_PhaseA', _cycle, outputDir=_output_dir, chosen_hue=_hue, _bins=c_bins)
SingleSaveHistPlot_bins(data_normal, 'Vabc_generator_CoefficientDistorsionContinue_PhaseB', _cycle, outputDir=_output_dir, chosen_hue=_hue, _bins=c_bins)
SingleSaveHistPlot_bins(data_normal, 'Vabc_generator_CoefficientDistorsionContinue_PhaseC', _cycle, outputDir=_output_dir, chosen_hue=_hue, _bins=c_bins)
SingleSaveHistPlot_bins(data_normal, 'Vabc_generator_DirectComponent_Value', _cycle, outputDir=_output_dir, chosen_hue=_hue, _bins=c_bins)
SingleSaveHistPlot_bins(data_normal, 'Vabc_generator_DirectComponent_Angle', _cycle, outputDir=_output_dir, chosen_hue=_hue, _bins=c_bins)
SingleSaveHistPlot_bins(data_normal, 'Vabc_generator_InverseComponent_Value', _cycle, outputDir=_output_dir, chosen_hue=_hue, _bins=c_bins)
SingleSaveHistPlot_bins(data_normal, 'Vabc_generator_InverseComponent_Angle', _cycle, outputDir=_output_dir, chosen_hue=_hue, _bins=c_bins)
SingleSaveHistPlot_bins(data_normal, 'Vabc_generator_ZeroComponent_Value', _cycle, outputDir=_output_dir, chosen_hue=_hue, _bins=c_bins)
SingleSaveHistPlot_bins(data_normal, 'Vabc_generator_ZeroComponent_Angle', _cycle, outputDir=_output_dir, chosen_hue=_hue, _bins=c_bins)
SingleSaveHistPlot_bins(data_normal, 'Iabc_loc_defect_ContinueComponent_PhaseA', _cycle, outputDir=_output_dir, chosen_hue=_hue, _bins=c_bins)
SingleSaveHistPlot_bins(data_normal, 'Iabc_loc_defect_ContinueComponent_PhaseB', _cycle, outputDir=_output_dir, chosen_hue=_hue, _bins=c_bins)
SingleSaveHistPlot_bins(data_normal, 'Iabc_loc_defect_ContinueComponent_PhaseC', _cycle, outputDir=_output_dir, chosen_hue=_hue, _bins=c_bins)
SingleSaveHistPlot_bins(data_normal, 'Iabc_loc_defect_EffectiveComponent_PhaseA', _cycle, outputDir=_output_dir, chosen_hue=_hue, _bins=c_bins)
SingleSaveHistPlot_bins(data_normal, 'Iabc_loc_defect_EffectiveComponent_PhaseB', _cycle, outputDir=_output_dir, chosen_hue=_hue, _bins=c_bins)
SingleSaveHistPlot_bins(data_normal, 'Iabc_loc_defect_EffectiveComponent_PhaseC', _cycle, outputDir=_output_dir, chosen_hue=_hue, _bins=c_bins)
SingleSaveHistPlot_bins(data_normal, 'Iabc_loc_defect_FirstArmonic_PhaseA_Value', _cycle, outputDir=_output_dir, chosen_hue=_hue, _bins=c_bins)
SingleSaveHistPlot_bins(data_normal, 'Iabc_loc_defect_FirstArmonic_PhaseA_Angle', _cycle, outputDir=_output_dir, chosen_hue=_hue, _bins=c_bins)
SingleSaveHistPlot_bins(data_normal, 'Iabc_loc_defect_FirstArmonic_PhaseB_Value', _cycle, outputDir=_output_dir, chosen_hue=_hue, _bins=c_bins)
SingleSaveHistPlot_bins(data_normal, 'Iabc_loc_defect_FirstArmonic_PhaseB_Angle', _cycle, outputDir=_output_dir, chosen_hue=_hue, _bins=c_bins)
SingleSaveHistPlot_bins(data_normal, 'Iabc_loc_defect_FirstArmonic_PhaseC_Value', _cycle, outputDir=_output_dir, chosen_hue=_hue, _bins=c_bins)
SingleSaveHistPlot_bins(data_normal, 'Iabc_loc_defect_FirstArmonic_PhaseC_Angle', _cycle, outputDir=_output_dir, chosen_hue=_hue, _bins=c_bins)
SingleSaveHistPlot_bins(data_normal, 'Iabc_loc_defect_DeformingComponent_PhaseA', _cycle, outputDir=_output_dir, chosen_hue=_hue, _bins=c_bins)
SingleSaveHistPlot_bins(data_normal, 'Iabc_loc_defect_DeformingComponent_PhaseB', _cycle, outputDir=_output_dir, chosen_hue=_hue, _bins=c_bins)
SingleSaveHistPlot_bins(data_normal, 'Iabc_loc_defect_DeformingComponent_PhaseC', _cycle, outputDir=_output_dir, chosen_hue=_hue, _bins=c_bins)
SingleSaveHistPlot_bins(data_normal, 'Iabc_loc_defect_CoefficientDistorsion_PhaseA', _cycle, outputDir=_output_dir, chosen_hue=_hue, _bins=c_bins)
SingleSaveHistPlot_bins(data_normal, 'Iabc_loc_defect_CoefficientDistorsion_PhaseB', _cycle, outputDir=_output_dir, chosen_hue=_hue, _bins=c_bins)
SingleSaveHistPlot_bins(data_normal, 'Iabc_loc_defect_CoefficientDistorsion_PhaseC', _cycle, outputDir=_output_dir, chosen_hue=_hue, _bins=c_bins)
SingleSaveHistPlot_bins(data_normal, 'Iabc_loc_defect_CoefficientDistorsionContinue_PhaseA', _cycle, outputDir=_output_dir, chosen_hue=_hue, _bins=c_bins)
SingleSaveHistPlot_bins(data_normal, 'Iabc_loc_defect_CoefficientDistorsionContinue_PhaseB', _cycle, outputDir=_output_dir, chosen_hue=_hue, _bins=c_bins)
SingleSaveHistPlot_bins(data_normal, 'Iabc_loc_defect_CoefficientDistorsionContinue_PhaseC', _cycle, outputDir=_output_dir, chosen_hue=_hue, _bins=c_bins)
SingleSaveHistPlot_bins(data_normal, 'Iabc_loc_defect_DirectComponent_Value', _cycle, outputDir=_output_dir, chosen_hue=_hue, _bins=c_bins)
SingleSaveHistPlot_bins(data_normal, 'Iabc_loc_defect_DirectComponent_Angle', _cycle, outputDir=_output_dir, chosen_hue=_hue, _bins=c_bins)
SingleSaveHistPlot_bins(data_normal, 'Iabc_loc_defect_InverseComponent_Value', _cycle, outputDir=_output_dir, chosen_hue=_hue, _bins=c_bins)
SingleSaveHistPlot_bins(data_normal, 'Iabc_loc_defect_InverseComponent_Angle', _cycle, outputDir=_output_dir, chosen_hue=_hue, _bins=c_bins)
SingleSaveHistPlot_bins(data_normal, 'Iabc_loc_defect_ZeroComponent_Value', _cycle, outputDir=_output_dir, chosen_hue=_hue, _bins=c_bins)
SingleSaveHistPlot_bins(data_normal, 'Iabc_loc_defect_ZeroComponent_Angle', _cycle, outputDir=_output_dir, chosen_hue=_hue, _bins=c_bins)
SingleSaveHistPlot_bins(data_normal, 'Vabc_loc_defect_ContinueComponent_PhaseA', _cycle, outputDir=_output_dir, chosen_hue=_hue, _bins=c_bins)
SingleSaveHistPlot_bins(data_normal, 'Vabc_loc_defect_ContinueComponent_PhaseB', _cycle, outputDir=_output_dir, chosen_hue=_hue, _bins=c_bins)
SingleSaveHistPlot_bins(data_normal, 'Vabc_loc_defect_ContinueComponent_PhaseC', _cycle, outputDir=_output_dir, chosen_hue=_hue, _bins=c_bins)
SingleSaveHistPlot_bins(data_normal, 'Vabc_loc_defect_EffectiveComponent_PhaseA', _cycle, outputDir=_output_dir, chosen_hue=_hue, _bins=c_bins)
SingleSaveHistPlot_bins(data_normal, 'Vabc_loc_defect_EffectiveComponent_PhaseB', _cycle, outputDir=_output_dir, chosen_hue=_hue, _bins=c_bins)
SingleSaveHistPlot_bins(data_normal, 'Vabc_loc_defect_EffectiveComponent_PhaseC', _cycle, outputDir=_output_dir, chosen_hue=_hue, _bins=c_bins)
SingleSaveHistPlot_bins(data_normal, 'Vabc_loc_defect_FirstArmonic_PhaseA_Value', _cycle, outputDir=_output_dir, chosen_hue=_hue, _bins=c_bins)
SingleSaveHistPlot_bins(data_normal, 'Vabc_loc_defect_FirstArmonic_PhaseA_Angle', _cycle, outputDir=_output_dir, chosen_hue=_hue, _bins=c_bins)
SingleSaveHistPlot_bins(data_normal, 'Vabc_loc_defect_FirstArmonic_PhaseB_Value', _cycle, outputDir=_output_dir, chosen_hue=_hue, _bins=c_bins)
SingleSaveHistPlot_bins(data_normal, 'Vabc_loc_defect_FirstArmonic_PhaseB_Angle', _cycle, outputDir=_output_dir, chosen_hue=_hue, _bins=c_bins)
SingleSaveHistPlot_bins(data_normal, 'Vabc_loc_defect_FirstArmonic_PhaseC_Value', _cycle, outputDir=_output_dir, chosen_hue=_hue, _bins=c_bins)
SingleSaveHistPlot_bins(data_normal, 'Vabc_loc_defect_FirstArmonic_PhaseC_Angle', _cycle, outputDir=_output_dir, chosen_hue=_hue, _bins=c_bins)
SingleSaveHistPlot_bins(data_normal, 'Vabc_loc_defect_DeformingComponent_PhaseA', _cycle, outputDir=_output_dir, chosen_hue=_hue, _bins=c_bins)
SingleSaveHistPlot_bins(data_normal, 'Vabc_loc_defect_DeformingComponent_PhaseB', _cycle, outputDir=_output_dir, chosen_hue=_hue, _bins=c_bins)
SingleSaveHistPlot_bins(data_normal, 'Vabc_loc_defect_DeformingComponent_PhaseC', _cycle, outputDir=_output_dir, chosen_hue=_hue, _bins=c_bins)
SingleSaveHistPlot_bins(data_normal, 'Vabc_loc_defect_CoefficientDistorsion_PhaseA', _cycle, outputDir=_output_dir, chosen_hue=_hue, _bins=c_bins)
SingleSaveHistPlot_bins(data_normal, 'Vabc_loc_defect_CoefficientDistorsion_PhaseB', _cycle, outputDir=_output_dir, chosen_hue=_hue, _bins=c_bins)
SingleSaveHistPlot_bins(data_normal, 'Vabc_loc_defect_CoefficientDistorsion_PhaseC', _cycle, outputDir=_output_dir, chosen_hue=_hue, _bins=c_bins)
SingleSaveHistPlot_bins(data_normal, 'Vabc_loc_defect_CoefficientDistorsionContinue_PhaseA', _cycle, outputDir=_output_dir, chosen_hue=_hue, _bins=c_bins)
SingleSaveHistPlot_bins(data_normal, 'Vabc_loc_defect_CoefficientDistorsionContinue_PhaseB', _cycle, outputDir=_output_dir, chosen_hue=_hue, _bins=c_bins)
SingleSaveHistPlot_bins(data_normal, 'Vabc_loc_defect_CoefficientDistorsionContinue_PhaseC', _cycle, outputDir=_output_dir, chosen_hue=_hue, _bins=c_bins)
SingleSaveHistPlot_bins(data_normal, 'Vabc_loc_defect_DirectComponent_Value', _cycle, outputDir=_output_dir, chosen_hue=_hue, _bins=c_bins)
SingleSaveHistPlot_bins(data_normal, 'Vabc_loc_defect_DirectComponent_Angle', _cycle, outputDir=_output_dir, chosen_hue=_hue, _bins=c_bins)
SingleSaveHistPlot_bins(data_normal, 'Vabc_loc_defect_InverseComponent_Value', _cycle, outputDir=_output_dir, chosen_hue=_hue, _bins=c_bins)
SingleSaveHistPlot_bins(data_normal, 'Vabc_loc_defect_InverseComponent_Angle', _cycle, outputDir=_output_dir, chosen_hue=_hue, _bins=c_bins)
SingleSaveHistPlot_bins(data_normal, 'Vabc_loc_defect_ZeroComponent_Value', _cycle, outputDir=_output_dir, chosen_hue=_hue, _bins=c_bins)
SingleSaveHistPlot_bins(data_normal, 'Vabc_loc_defect_ZeroComponent_Angle', _cycle, outputDir=_output_dir, chosen_hue=_hue, _bins=c_bins)
SingleSaveHistPlot_bins(data_normal, 'Iabc_consumator_ContinueComponent_PhaseA', _cycle, outputDir=_output_dir, chosen_hue=_hue, _bins=c_bins)
SingleSaveHistPlot_bins(data_normal, 'Iabc_consumator_ContinueComponent_PhaseB', _cycle, outputDir=_output_dir, chosen_hue=_hue, _bins=c_bins)
SingleSaveHistPlot_bins(data_normal, 'Iabc_consumator_ContinueComponent_PhaseC', _cycle, outputDir=_output_dir, chosen_hue=_hue, _bins=c_bins)
SingleSaveHistPlot_bins(data_normal, 'Iabc_consumator_EffectiveComponent_PhaseA', _cycle, outputDir=_output_dir, chosen_hue=_hue, _bins=c_bins)
SingleSaveHistPlot_bins(data_normal, 'Iabc_consumator_EffectiveComponent_PhaseB', _cycle, outputDir=_output_dir, chosen_hue=_hue, _bins=c_bins)
SingleSaveHistPlot_bins(data_normal, 'Iabc_consumator_EffectiveComponent_PhaseC', _cycle, outputDir=_output_dir, chosen_hue=_hue, _bins=c_bins)
SingleSaveHistPlot_bins(data_normal, 'Iabc_consumator_FirstArmonic_PhaseA_Value', _cycle, outputDir=_output_dir, chosen_hue=_hue, _bins=c_bins)
SingleSaveHistPlot_bins(data_normal, 'Iabc_consumator_FirstArmonic_PhaseA_Angle', _cycle, outputDir=_output_dir, chosen_hue=_hue, _bins=c_bins)
SingleSaveHistPlot_bins(data_normal, 'Iabc_consumator_FirstArmonic_PhaseB_Value', _cycle, outputDir=_output_dir, chosen_hue=_hue, _bins=c_bins)
SingleSaveHistPlot_bins(data_normal, 'Iabc_consumator_FirstArmonic_PhaseB_Angle', _cycle, outputDir=_output_dir, chosen_hue=_hue, _bins=c_bins)
SingleSaveHistPlot_bins(data_normal, 'Iabc_consumator_FirstArmonic_PhaseC_Value', _cycle, outputDir=_output_dir, chosen_hue=_hue, _bins=c_bins)
SingleSaveHistPlot_bins(data_normal, 'Iabc_consumator_FirstArmonic_PhaseC_Angle', _cycle, outputDir=_output_dir, chosen_hue=_hue, _bins=c_bins)
SingleSaveHistPlot_bins(data_normal, 'Iabc_consumator_DeformingComponent_PhaseA', _cycle, outputDir=_output_dir, chosen_hue=_hue, _bins=c_bins)
SingleSaveHistPlot_bins(data_normal, 'Iabc_consumator_DeformingComponent_PhaseB', _cycle, outputDir=_output_dir, chosen_hue=_hue, _bins=c_bins)
SingleSaveHistPlot_bins(data_normal, 'Iabc_consumator_DeformingComponent_PhaseC', _cycle, outputDir=_output_dir, chosen_hue=_hue, _bins=c_bins)
SingleSaveHistPlot_bins(data_normal, 'Iabc_consumator_CoefficientDistorsion_PhaseA', _cycle, outputDir=_output_dir, chosen_hue=_hue, _bins=c_bins)
SingleSaveHistPlot_bins(data_normal, 'Iabc_consumator_CoefficientDistorsion_PhaseB', _cycle, outputDir=_output_dir, chosen_hue=_hue, _bins=c_bins)
SingleSaveHistPlot_bins(data_normal, 'Iabc_consumator_CoefficientDistorsion_PhaseC', _cycle, outputDir=_output_dir, chosen_hue=_hue, _bins=c_bins)
SingleSaveHistPlot_bins(data_normal, 'Iabc_consumator_CoefficientDistorsionContinue_PhaseA', _cycle, outputDir=_output_dir, chosen_hue=_hue, _bins=c_bins)
SingleSaveHistPlot_bins(data_normal, 'Iabc_consumator_CoefficientDistorsionContinue_PhaseB', _cycle, outputDir=_output_dir, chosen_hue=_hue, _bins=c_bins)
SingleSaveHistPlot_bins(data_normal, 'Iabc_consumator_CoefficientDistorsionContinue_PhaseC', _cycle, outputDir=_output_dir, chosen_hue=_hue, _bins=c_bins)
SingleSaveHistPlot_bins(data_normal, 'Iabc_consumator_DirectComponent_Value', _cycle, outputDir=_output_dir, chosen_hue=_hue, _bins=c_bins)
SingleSaveHistPlot_bins(data_normal, 'Iabc_consumator_DirectComponent_Angle', _cycle, outputDir=_output_dir, chosen_hue=_hue, _bins=c_bins)
SingleSaveHistPlot_bins(data_normal, 'Iabc_consumator_InverseComponent_Value', _cycle, outputDir=_output_dir, chosen_hue=_hue, _bins=c_bins)
SingleSaveHistPlot_bins(data_normal, 'Iabc_consumator_InverseComponent_Angle', _cycle, outputDir=_output_dir, chosen_hue=_hue, _bins=c_bins)
SingleSaveHistPlot_bins(data_normal, 'Iabc_consumator_ZeroComponent_Value', _cycle, outputDir=_output_dir, chosen_hue=_hue, _bins=c_bins)
SingleSaveHistPlot_bins(data_normal, 'Iabc_consumator_ZeroComponent_Angle', _cycle, outputDir=_output_dir, chosen_hue=_hue, _bins=c_bins)
SingleSaveHistPlot_bins(data_normal, 'Vabc_consumator_ContinueComponent_PhaseA', _cycle, outputDir=_output_dir, chosen_hue=_hue, _bins=c_bins)
SingleSaveHistPlot_bins(data_normal, 'Vabc_consumator_ContinueComponent_PhaseB', _cycle, outputDir=_output_dir, chosen_hue=_hue, _bins=c_bins)
SingleSaveHistPlot_bins(data_normal, 'Vabc_consumator_ContinueComponent_PhaseC', _cycle, outputDir=_output_dir, chosen_hue=_hue, _bins=c_bins)
SingleSaveHistPlot_bins(data_normal, 'Vabc_consumator_EffectiveComponent_PhaseA', _cycle, outputDir=_output_dir, chosen_hue=_hue, _bins=c_bins)
SingleSaveHistPlot_bins(data_normal, 'Vabc_consumator_EffectiveComponent_PhaseB', _cycle, outputDir=_output_dir, chosen_hue=_hue, _bins=c_bins)
SingleSaveHistPlot_bins(data_normal, 'Vabc_consumator_EffectiveComponent_PhaseC', _cycle, outputDir=_output_dir, chosen_hue=_hue, _bins=c_bins)
SingleSaveHistPlot_bins(data_normal, 'Vabc_consumator_FirstArmonic_PhaseA_Value', _cycle, outputDir=_output_dir, chosen_hue=_hue, _bins=c_bins)
SingleSaveHistPlot_bins(data_normal, 'Vabc_consumator_FirstArmonic_PhaseA_Angle', _cycle, outputDir=_output_dir, chosen_hue=_hue, _bins=c_bins)
SingleSaveHistPlot_bins(data_normal, 'Vabc_consumator_FirstArmonic_PhaseB_Value', _cycle, outputDir=_output_dir, chosen_hue=_hue, _bins=c_bins)
SingleSaveHistPlot_bins(data_normal, 'Vabc_consumator_FirstArmonic_PhaseB_Angle', _cycle, outputDir=_output_dir, chosen_hue=_hue, _bins=c_bins)
SingleSaveHistPlot_bins(data_normal, 'Vabc_consumator_FirstArmonic_PhaseC_Value', _cycle, outputDir=_output_dir, chosen_hue=_hue, _bins=c_bins)
SingleSaveHistPlot_bins(data_normal, 'Vabc_consumator_FirstArmonic_PhaseC_Angle', _cycle, outputDir=_output_dir, chosen_hue=_hue, _bins=c_bins)
SingleSaveHistPlot_bins(data_normal, 'Vabc_consumator_DeformingComponent_PhaseA', _cycle, outputDir=_output_dir, chosen_hue=_hue, _bins=c_bins)
SingleSaveHistPlot_bins(data_normal, 'Vabc_consumator_DeformingComponent_PhaseB', _cycle, outputDir=_output_dir, chosen_hue=_hue, _bins=c_bins)
SingleSaveHistPlot_bins(data_normal, 'Vabc_consumator_DeformingComponent_PhaseC', _cycle, outputDir=_output_dir, chosen_hue=_hue, _bins=c_bins)
SingleSaveHistPlot_bins(data_normal, 'Vabc_consumator_CoefficientDistorsion_PhaseA', _cycle, outputDir=_output_dir, chosen_hue=_hue, _bins=c_bins)
SingleSaveHistPlot_bins(data_normal, 'Vabc_consumator_CoefficientDistorsion_PhaseB', _cycle, outputDir=_output_dir, chosen_hue=_hue, _bins=c_bins)
SingleSaveHistPlot_bins(data_normal, 'Vabc_consumator_CoefficientDistorsion_PhaseC', _cycle, outputDir=_output_dir, chosen_hue=_hue, _bins=c_bins)
SingleSaveHistPlot_bins(data_normal, 'Vabc_consumator_CoefficientDistorsionContinue_PhaseA', _cycle, outputDir=_output_dir, chosen_hue=_hue, _bins=c_bins)
SingleSaveHistPlot_bins(data_normal, 'Vabc_consumator_CoefficientDistorsionContinue_PhaseB', _cycle, outputDir=_output_dir, chosen_hue=_hue, _bins=c_bins)
SingleSaveHistPlot_bins(data_normal, 'Vabc_consumator_CoefficientDistorsionContinue_PhaseC', _cycle, outputDir=_output_dir, chosen_hue=_hue, _bins=c_bins)
SingleSaveHistPlot_bins(data_normal, 'Vabc_consumator_DirectComponent_Value', _cycle, outputDir=_output_dir, chosen_hue=_hue, _bins=c_bins)
SingleSaveHistPlot_bins(data_normal, 'Vabc_consumator_DirectComponent_Angle', _cycle, outputDir=_output_dir, chosen_hue=_hue, _bins=c_bins)
SingleSaveHistPlot_bins(data_normal, 'Vabc_consumator_InverseComponent_Value', _cycle, outputDir=_output_dir, chosen_hue=_hue, _bins=c_bins)
SingleSaveHistPlot_bins(data_normal, 'Vabc_consumator_InverseComponent_Angle', _cycle, outputDir=_output_dir, chosen_hue=_hue, _bins=c_bins)
SingleSaveHistPlot_bins(data_normal, 'Vabc_consumator_ZeroComponent_Value', _cycle, outputDir=_output_dir, chosen_hue=_hue, _bins=c_bins)
SingleSaveHistPlot_bins(data_normal, 'Vabc_consumator_ZeroComponent_Angle', _cycle, outputDir=_output_dir, chosen_hue=_hue, _bins=c_bins)


SingleSaveHistPlot_bins(data_normal, 'Iabc_generator_CoefficientDistorsionFirstArmonic_PhaseA', _cycle, outputDir=_output_dir, chosen_hue=_hue, _bins=c_bins)
SingleSaveHistPlot_bins(data_normal, 'Iabc_generator_CoefficientDistorsionFirstArmonic_PhaseB', _cycle, outputDir=_output_dir, chosen_hue=_hue, _bins=c_bins)
SingleSaveHistPlot_bins(data_normal, 'Iabc_generator_CoefficientDistorsionFirstArmonic_PhaseC', _cycle, outputDir=_output_dir, chosen_hue=_hue, _bins=c_bins)
SingleSaveHistPlot_bins(data_normal, 'Iabc_generator_CoefficeintNegativeAsimetry', _cycle, outputDir=_output_dir, chosen_hue=_hue, _bins=c_bins)
SingleSaveHistPlot_bins(data_normal, 'Iabc_generator_CoefficeintZeroAsimetry', _cycle, outputDir=_output_dir, chosen_hue=_hue, _bins=c_bins)
SingleSaveHistPlot_bins(data_normal, 'Vabc_generator_CoefficientDistorsionFirstArmonic_PhaseA', _cycle, outputDir=_output_dir, chosen_hue=_hue, _bins=c_bins)
SingleSaveHistPlot_bins(data_normal, 'Vabc_generator_CoefficientDistorsionFirstArmonic_PhaseB', _cycle, outputDir=_output_dir, chosen_hue=_hue, _bins=c_bins)
SingleSaveHistPlot_bins(data_normal, 'Vabc_generator_CoefficientDistorsionFirstArmonic_PhaseC', _cycle, outputDir=_output_dir, chosen_hue=_hue, _bins=c_bins)
SingleSaveHistPlot_bins(data_normal, 'Vabc_generator_CoefficeintNegativeAsimetry', _cycle, outputDir=_output_dir, chosen_hue=_hue, _bins=c_bins)
SingleSaveHistPlot_bins(data_normal, 'Vabc_generator_CoefficeintZeroAsimetry', _cycle, outputDir=_output_dir, chosen_hue=_hue, _bins=c_bins)
SingleSaveHistPlot_bins(data_normal, 'Iabc_loc_defect_CoefficientDistorsionFirstArmonic_PhaseA', _cycle, outputDir=_output_dir, chosen_hue=_hue, _bins=c_bins)
SingleSaveHistPlot_bins(data_normal, 'Iabc_loc_defect_CoefficientDistorsionFirstArmonic_PhaseB', _cycle, outputDir=_output_dir, chosen_hue=_hue, _bins=c_bins)
SingleSaveHistPlot_bins(data_normal, 'Iabc_loc_defect_CoefficientDistorsionFirstArmonic_PhaseC', _cycle, outputDir=_output_dir, chosen_hue=_hue, _bins=c_bins)
SingleSaveHistPlot_bins(data_normal, 'Iabc_loc_defect_CoefficeintNegativeAsimetry', _cycle, outputDir=_output_dir, chosen_hue=_hue, _bins=c_bins)
SingleSaveHistPlot_bins(data_normal, 'Iabc_loc_defect_CoefficeintZeroAsimetry', _cycle, outputDir=_output_dir, chosen_hue=_hue, _bins=c_bins)
SingleSaveHistPlot_bins(data_normal, 'Vabc_loc_defect_CoefficientDistorsionFirstArmonic_PhaseA', _cycle, outputDir=_output_dir, chosen_hue=_hue, _bins=c_bins)
SingleSaveHistPlot_bins(data_normal, 'Vabc_loc_defect_CoefficientDistorsionFirstArmonic_PhaseB', _cycle, outputDir=_output_dir, chosen_hue=_hue, _bins=c_bins)
SingleSaveHistPlot_bins(data_normal, 'Vabc_loc_defect_CoefficientDistorsionFirstArmonic_PhaseC', _cycle, outputDir=_output_dir, chosen_hue=_hue, _bins=c_bins)
SingleSaveHistPlot_bins(data_normal, 'Vabc_loc_defect_CoefficeintNegativeAsimetry', _cycle, outputDir=_output_dir, chosen_hue=_hue, _bins=c_bins)
SingleSaveHistPlot_bins(data_normal, 'Vabc_loc_defect_CoefficeintZeroAsimetry', _cycle, outputDir=_output_dir, chosen_hue=_hue, _bins=c_bins)
SingleSaveHistPlot_bins(data_normal, 'Iabc_consumator_CoefficientDistorsionFirstArmonic_PhaseA', _cycle, outputDir=_output_dir, chosen_hue=_hue, _bins=c_bins)
SingleSaveHistPlot_bins(data_normal, 'Iabc_consumator_CoefficientDistorsionFirstArmonic_PhaseB', _cycle, outputDir=_output_dir, chosen_hue=_hue, _bins=c_bins)
SingleSaveHistPlot_bins(data_normal, 'Iabc_consumator_CoefficientDistorsionFirstArmonic_PhaseC', _cycle, outputDir=_output_dir, chosen_hue=_hue, _bins=c_bins)
SingleSaveHistPlot_bins(data_normal, 'Iabc_consumator_CoefficeintNegativeAsimetry', _cycle, outputDir=_output_dir, chosen_hue=_hue, _bins=c_bins)
SingleSaveHistPlot_bins(data_normal, 'Iabc_consumator_CoefficeintZeroAsimetry', _cycle, outputDir=_output_dir, chosen_hue=_hue, _bins=c_bins)
SingleSaveHistPlot_bins(data_normal, 'Vabc_consumator_CoefficientDistorsionFirstArmonic_PhaseA', _cycle, outputDir=_output_dir, chosen_hue=_hue, _bins=c_bins)
SingleSaveHistPlot_bins(data_normal, 'Vabc_consumator_CoefficientDistorsionFirstArmonic_PhaseB', _cycle, outputDir=_output_dir, chosen_hue=_hue, _bins=c_bins)
SingleSaveHistPlot_bins(data_normal, 'Vabc_consumator_CoefficientDistorsionFirstArmonic_PhaseC', _cycle, outputDir=_output_dir, chosen_hue=_hue, _bins=c_bins)
SingleSaveHistPlot_bins(data_normal, 'Vabc_consumator_CoefficeintNegativeAsimetry', _cycle, outputDir=_output_dir, chosen_hue=_hue, _bins=c_bins)
SingleSaveHistPlot_bins(data_normal, 'Vabc_consumator_CoefficeintZeroAsimetry', _cycle, outputDir=_output_dir, chosen_hue=_hue, _bins=c_bins)










_output_dir="data_normal_Phase"
_hue='phase'

_cycle=19
SingleSaveHistPlot(data_normal, 'Iabc_generator_ContinueComponent_PhaseA', _cycle, outputDir=_output_dir, chosen_hue=_hue)
SingleSaveHistPlot(data_normal, 'Iabc_generator_ContinueComponent_PhaseB', _cycle, outputDir=_output_dir, chosen_hue=_hue)
SingleSaveHistPlot(data_normal, 'Iabc_generator_ContinueComponent_PhaseC', _cycle, outputDir=_output_dir, chosen_hue=_hue)
SingleSaveHistPlot(data_normal, 'Iabc_generator_EffectiveComponent_PhaseA', _cycle, outputDir=_output_dir, chosen_hue=_hue)
SingleSaveHistPlot(data_normal, 'Iabc_generator_EffectiveComponent_PhaseB', _cycle, outputDir=_output_dir, chosen_hue=_hue)
SingleSaveHistPlot(data_normal, 'Iabc_generator_EffectiveComponent_PhaseC', _cycle, outputDir=_output_dir, chosen_hue=_hue)
SingleSaveHistPlot(data_normal, 'Iabc_generator_FirstArmonic_PhaseA_Value', _cycle, outputDir=_output_dir, chosen_hue=_hue)
SingleSaveHistPlot(data_normal, 'Iabc_generator_FirstArmonic_PhaseA_Angle', _cycle, outputDir=_output_dir, chosen_hue=_hue)
SingleSaveHistPlot(data_normal, 'Iabc_generator_FirstArmonic_PhaseB_Value', _cycle, outputDir=_output_dir, chosen_hue=_hue)
SingleSaveHistPlot(data_normal, 'Iabc_generator_FirstArmonic_PhaseB_Angle', _cycle, outputDir=_output_dir, chosen_hue=_hue)
SingleSaveHistPlot(data_normal, 'Iabc_generator_FirstArmonic_PhaseC_Value', _cycle, outputDir=_output_dir, chosen_hue=_hue)
SingleSaveHistPlot(data_normal, 'Iabc_generator_FirstArmonic_PhaseC_Angle', _cycle, outputDir=_output_dir, chosen_hue=_hue)
SingleSaveHistPlot(data_normal, 'Iabc_generator_DeformingComponent_PhaseA', _cycle, outputDir=_output_dir, chosen_hue=_hue)
SingleSaveHistPlot(data_normal, 'Iabc_generator_DeformingComponent_PhaseB', _cycle, outputDir=_output_dir, chosen_hue=_hue)
SingleSaveHistPlot(data_normal, 'Iabc_generator_DeformingComponent_PhaseC', _cycle, outputDir=_output_dir, chosen_hue=_hue)
SingleSaveHistPlot(data_normal, 'Iabc_generator_CoefficientDistorsion_PhaseA', _cycle, outputDir=_output_dir, chosen_hue=_hue)
SingleSaveHistPlot(data_normal, 'Iabc_generator_CoefficientDistorsion_PhaseB', _cycle, outputDir=_output_dir, chosen_hue=_hue)
SingleSaveHistPlot(data_normal, 'Iabc_generator_CoefficientDistorsion_PhaseC', _cycle, outputDir=_output_dir, chosen_hue=_hue)
SingleSaveHistPlot(data_normal, 'Iabc_generator_CoefficientDistorsionContinue_PhaseA', _cycle, outputDir=_output_dir, chosen_hue=_hue)
SingleSaveHistPlot(data_normal, 'Iabc_generator_CoefficientDistorsionContinue_PhaseB', _cycle, outputDir=_output_dir, chosen_hue=_hue)
SingleSaveHistPlot(data_normal, 'Iabc_generator_CoefficientDistorsionContinue_PhaseC', _cycle, outputDir=_output_dir, chosen_hue=_hue)
SingleSaveHistPlot(data_normal, 'Iabc_generator_DirectComponent_Value', _cycle, outputDir=_output_dir, chosen_hue=_hue)
SingleSaveHistPlot(data_normal, 'Iabc_generator_DirectComponent_Angle', _cycle, outputDir=_output_dir, chosen_hue=_hue)
SingleSaveHistPlot(data_normal, 'Iabc_generator_InverseComponent_Value', _cycle, outputDir=_output_dir, chosen_hue=_hue)
SingleSaveHistPlot(data_normal, 'Iabc_generator_InverseComponent_Angle', _cycle, outputDir=_output_dir, chosen_hue=_hue)
SingleSaveHistPlot(data_normal, 'Iabc_generator_ZeroComponent_Value', _cycle, outputDir=_output_dir, chosen_hue=_hue)
SingleSaveHistPlot(data_normal, 'Iabc_generator_ZeroComponent_Angle', _cycle, outputDir=_output_dir, chosen_hue=_hue)
SingleSaveHistPlot(data_normal, 'Vabc_generator_ContinueComponent_PhaseA', _cycle, outputDir=_output_dir, chosen_hue=_hue)
SingleSaveHistPlot(data_normal, 'Vabc_generator_ContinueComponent_PhaseB', _cycle, outputDir=_output_dir, chosen_hue=_hue)
SingleSaveHistPlot(data_normal, 'Vabc_generator_ContinueComponent_PhaseC', _cycle, outputDir=_output_dir, chosen_hue=_hue)
SingleSaveHistPlot(data_normal, 'Vabc_generator_EffectiveComponent_PhaseA', _cycle, outputDir=_output_dir, chosen_hue=_hue)
SingleSaveHistPlot(data_normal, 'Vabc_generator_EffectiveComponent_PhaseB', _cycle, outputDir=_output_dir, chosen_hue=_hue)
SingleSaveHistPlot(data_normal, 'Vabc_generator_EffectiveComponent_PhaseC', _cycle, outputDir=_output_dir, chosen_hue=_hue)
SingleSaveHistPlot(data_normal, 'Vabc_generator_FirstArmonic_PhaseA_Value', _cycle, outputDir=_output_dir, chosen_hue=_hue)
SingleSaveHistPlot(data_normal, 'Vabc_generator_FirstArmonic_PhaseA_Angle', _cycle, outputDir=_output_dir, chosen_hue=_hue)
SingleSaveHistPlot(data_normal, 'Vabc_generator_FirstArmonic_PhaseB_Value', _cycle, outputDir=_output_dir, chosen_hue=_hue)
SingleSaveHistPlot(data_normal, 'Vabc_generator_FirstArmonic_PhaseB_Angle', _cycle, outputDir=_output_dir, chosen_hue=_hue)
SingleSaveHistPlot(data_normal, 'Vabc_generator_FirstArmonic_PhaseC_Value', _cycle, outputDir=_output_dir, chosen_hue=_hue)
SingleSaveHistPlot(data_normal, 'Vabc_generator_FirstArmonic_PhaseC_Angle', _cycle, outputDir=_output_dir, chosen_hue=_hue)
SingleSaveHistPlot(data_normal, 'Vabc_generator_DeformingComponent_PhaseA', _cycle, outputDir=_output_dir, chosen_hue=_hue)
SingleSaveHistPlot(data_normal, 'Vabc_generator_DeformingComponent_PhaseB', _cycle, outputDir=_output_dir, chosen_hue=_hue)
SingleSaveHistPlot(data_normal, 'Vabc_generator_DeformingComponent_PhaseC', _cycle, outputDir=_output_dir, chosen_hue=_hue)
SingleSaveHistPlot(data_normal, 'Vabc_generator_CoefficientDistorsion_PhaseA', _cycle, outputDir=_output_dir, chosen_hue=_hue)
SingleSaveHistPlot(data_normal, 'Vabc_generator_CoefficientDistorsion_PhaseB', _cycle, outputDir=_output_dir, chosen_hue=_hue)
SingleSaveHistPlot(data_normal, 'Vabc_generator_CoefficientDistorsion_PhaseC', _cycle, outputDir=_output_dir, chosen_hue=_hue)
SingleSaveHistPlot(data_normal, 'Vabc_generator_CoefficientDistorsionContinue_PhaseA', _cycle, outputDir=_output_dir, chosen_hue=_hue)
SingleSaveHistPlot(data_normal, 'Vabc_generator_CoefficientDistorsionContinue_PhaseB', _cycle, outputDir=_output_dir, chosen_hue=_hue)
SingleSaveHistPlot(data_normal, 'Vabc_generator_CoefficientDistorsionContinue_PhaseC', _cycle, outputDir=_output_dir, chosen_hue=_hue)
SingleSaveHistPlot(data_normal, 'Vabc_generator_DirectComponent_Value', _cycle, outputDir=_output_dir, chosen_hue=_hue)
SingleSaveHistPlot(data_normal, 'Vabc_generator_DirectComponent_Angle', _cycle, outputDir=_output_dir, chosen_hue=_hue)
SingleSaveHistPlot(data_normal, 'Vabc_generator_InverseComponent_Value', _cycle, outputDir=_output_dir, chosen_hue=_hue)
SingleSaveHistPlot(data_normal, 'Vabc_generator_InverseComponent_Angle', _cycle, outputDir=_output_dir, chosen_hue=_hue)
SingleSaveHistPlot(data_normal, 'Vabc_generator_ZeroComponent_Value', _cycle, outputDir=_output_dir, chosen_hue=_hue)
SingleSaveHistPlot(data_normal, 'Vabc_generator_ZeroComponent_Angle', _cycle, outputDir=_output_dir, chosen_hue=_hue)
SingleSaveHistPlot(data_normal, 'Iabc_loc_defect_ContinueComponent_PhaseA', _cycle, outputDir=_output_dir, chosen_hue=_hue)
SingleSaveHistPlot(data_normal, 'Iabc_loc_defect_ContinueComponent_PhaseB', _cycle, outputDir=_output_dir, chosen_hue=_hue)
SingleSaveHistPlot(data_normal, 'Iabc_loc_defect_ContinueComponent_PhaseC', _cycle, outputDir=_output_dir, chosen_hue=_hue)
SingleSaveHistPlot(data_normal, 'Iabc_loc_defect_EffectiveComponent_PhaseA', _cycle, outputDir=_output_dir, chosen_hue=_hue)
SingleSaveHistPlot(data_normal, 'Iabc_loc_defect_EffectiveComponent_PhaseB', _cycle, outputDir=_output_dir, chosen_hue=_hue)
SingleSaveHistPlot(data_normal, 'Iabc_loc_defect_EffectiveComponent_PhaseC', _cycle, outputDir=_output_dir, chosen_hue=_hue)
SingleSaveHistPlot(data_normal, 'Iabc_loc_defect_FirstArmonic_PhaseA_Value', _cycle, outputDir=_output_dir, chosen_hue=_hue)
SingleSaveHistPlot(data_normal, 'Iabc_loc_defect_FirstArmonic_PhaseA_Angle', _cycle, outputDir=_output_dir, chosen_hue=_hue)
SingleSaveHistPlot(data_normal, 'Iabc_loc_defect_FirstArmonic_PhaseB_Value', _cycle, outputDir=_output_dir, chosen_hue=_hue)
SingleSaveHistPlot(data_normal, 'Iabc_loc_defect_FirstArmonic_PhaseB_Angle', _cycle, outputDir=_output_dir, chosen_hue=_hue)
SingleSaveHistPlot(data_normal, 'Iabc_loc_defect_FirstArmonic_PhaseC_Value', _cycle, outputDir=_output_dir, chosen_hue=_hue)
SingleSaveHistPlot(data_normal, 'Iabc_loc_defect_FirstArmonic_PhaseC_Angle', _cycle, outputDir=_output_dir, chosen_hue=_hue)
SingleSaveHistPlot(data_normal, 'Iabc_loc_defect_DeformingComponent_PhaseA', _cycle, outputDir=_output_dir, chosen_hue=_hue)
SingleSaveHistPlot(data_normal, 'Iabc_loc_defect_DeformingComponent_PhaseB', _cycle, outputDir=_output_dir, chosen_hue=_hue)
SingleSaveHistPlot(data_normal, 'Iabc_loc_defect_DeformingComponent_PhaseC', _cycle, outputDir=_output_dir, chosen_hue=_hue)
SingleSaveHistPlot(data_normal, 'Iabc_loc_defect_CoefficientDistorsion_PhaseA', _cycle, outputDir=_output_dir, chosen_hue=_hue)
SingleSaveHistPlot(data_normal, 'Iabc_loc_defect_CoefficientDistorsion_PhaseB', _cycle, outputDir=_output_dir, chosen_hue=_hue)
SingleSaveHistPlot(data_normal, 'Iabc_loc_defect_CoefficientDistorsion_PhaseC', _cycle, outputDir=_output_dir, chosen_hue=_hue)
SingleSaveHistPlot(data_normal, 'Iabc_loc_defect_CoefficientDistorsionContinue_PhaseA', _cycle, outputDir=_output_dir, chosen_hue=_hue)
SingleSaveHistPlot(data_normal, 'Iabc_loc_defect_CoefficientDistorsionContinue_PhaseB', _cycle, outputDir=_output_dir, chosen_hue=_hue)
SingleSaveHistPlot(data_normal, 'Iabc_loc_defect_CoefficientDistorsionContinue_PhaseC', _cycle, outputDir=_output_dir, chosen_hue=_hue)
SingleSaveHistPlot(data_normal, 'Iabc_loc_defect_DirectComponent_Value', _cycle, outputDir=_output_dir, chosen_hue=_hue)
SingleSaveHistPlot(data_normal, 'Iabc_loc_defect_DirectComponent_Angle', _cycle, outputDir=_output_dir, chosen_hue=_hue)
SingleSaveHistPlot(data_normal, 'Iabc_loc_defect_InverseComponent_Value', _cycle, outputDir=_output_dir, chosen_hue=_hue)
SingleSaveHistPlot(data_normal, 'Iabc_loc_defect_InverseComponent_Angle', _cycle, outputDir=_output_dir, chosen_hue=_hue)
SingleSaveHistPlot(data_normal, 'Iabc_loc_defect_ZeroComponent_Value', _cycle, outputDir=_output_dir, chosen_hue=_hue)
SingleSaveHistPlot(data_normal, 'Iabc_loc_defect_ZeroComponent_Angle', _cycle, outputDir=_output_dir, chosen_hue=_hue)
SingleSaveHistPlot(data_normal, 'Vabc_loc_defect_ContinueComponent_PhaseA', _cycle, outputDir=_output_dir, chosen_hue=_hue)
SingleSaveHistPlot(data_normal, 'Vabc_loc_defect_ContinueComponent_PhaseB', _cycle, outputDir=_output_dir, chosen_hue=_hue)
SingleSaveHistPlot(data_normal, 'Vabc_loc_defect_ContinueComponent_PhaseC', _cycle, outputDir=_output_dir, chosen_hue=_hue)
SingleSaveHistPlot(data_normal, 'Vabc_loc_defect_EffectiveComponent_PhaseA', _cycle, outputDir=_output_dir, chosen_hue=_hue)
SingleSaveHistPlot(data_normal, 'Vabc_loc_defect_EffectiveComponent_PhaseB', _cycle, outputDir=_output_dir, chosen_hue=_hue)
SingleSaveHistPlot(data_normal, 'Vabc_loc_defect_EffectiveComponent_PhaseC', _cycle, outputDir=_output_dir, chosen_hue=_hue)
SingleSaveHistPlot(data_normal, 'Vabc_loc_defect_FirstArmonic_PhaseA_Value', _cycle, outputDir=_output_dir, chosen_hue=_hue)
SingleSaveHistPlot(data_normal, 'Vabc_loc_defect_FirstArmonic_PhaseA_Angle', _cycle, outputDir=_output_dir, chosen_hue=_hue)
SingleSaveHistPlot(data_normal, 'Vabc_loc_defect_FirstArmonic_PhaseB_Value', _cycle, outputDir=_output_dir, chosen_hue=_hue)
SingleSaveHistPlot(data_normal, 'Vabc_loc_defect_FirstArmonic_PhaseB_Angle', _cycle, outputDir=_output_dir, chosen_hue=_hue)
SingleSaveHistPlot(data_normal, 'Vabc_loc_defect_FirstArmonic_PhaseC_Value', _cycle, outputDir=_output_dir, chosen_hue=_hue)
SingleSaveHistPlot(data_normal, 'Vabc_loc_defect_FirstArmonic_PhaseC_Angle', _cycle, outputDir=_output_dir, chosen_hue=_hue)
SingleSaveHistPlot(data_normal, 'Vabc_loc_defect_DeformingComponent_PhaseA', _cycle, outputDir=_output_dir, chosen_hue=_hue)
SingleSaveHistPlot(data_normal, 'Vabc_loc_defect_DeformingComponent_PhaseB', _cycle, outputDir=_output_dir, chosen_hue=_hue)
SingleSaveHistPlot(data_normal, 'Vabc_loc_defect_DeformingComponent_PhaseC', _cycle, outputDir=_output_dir, chosen_hue=_hue)
SingleSaveHistPlot(data_normal, 'Vabc_loc_defect_CoefficientDistorsion_PhaseA', _cycle, outputDir=_output_dir, chosen_hue=_hue)
SingleSaveHistPlot(data_normal, 'Vabc_loc_defect_CoefficientDistorsion_PhaseB', _cycle, outputDir=_output_dir, chosen_hue=_hue)
SingleSaveHistPlot(data_normal, 'Vabc_loc_defect_CoefficientDistorsion_PhaseC', _cycle, outputDir=_output_dir, chosen_hue=_hue)
SingleSaveHistPlot(data_normal, 'Vabc_loc_defect_CoefficientDistorsionContinue_PhaseA', _cycle, outputDir=_output_dir, chosen_hue=_hue)
SingleSaveHistPlot(data_normal, 'Vabc_loc_defect_CoefficientDistorsionContinue_PhaseB', _cycle, outputDir=_output_dir, chosen_hue=_hue)
SingleSaveHistPlot(data_normal, 'Vabc_loc_defect_CoefficientDistorsionContinue_PhaseC', _cycle, outputDir=_output_dir, chosen_hue=_hue)
SingleSaveHistPlot(data_normal, 'Vabc_loc_defect_DirectComponent_Value', _cycle, outputDir=_output_dir, chosen_hue=_hue)
SingleSaveHistPlot(data_normal, 'Vabc_loc_defect_DirectComponent_Angle', _cycle, outputDir=_output_dir, chosen_hue=_hue)
SingleSaveHistPlot(data_normal, 'Vabc_loc_defect_InverseComponent_Value', _cycle, outputDir=_output_dir, chosen_hue=_hue)
SingleSaveHistPlot(data_normal, 'Vabc_loc_defect_InverseComponent_Angle', _cycle, outputDir=_output_dir, chosen_hue=_hue)
SingleSaveHistPlot(data_normal, 'Vabc_loc_defect_ZeroComponent_Value', _cycle, outputDir=_output_dir, chosen_hue=_hue)
SingleSaveHistPlot(data_normal, 'Vabc_loc_defect_ZeroComponent_Angle', _cycle, outputDir=_output_dir, chosen_hue=_hue)
SingleSaveHistPlot(data_normal, 'Iabc_consumator_ContinueComponent_PhaseA', _cycle, outputDir=_output_dir, chosen_hue=_hue)
SingleSaveHistPlot(data_normal, 'Iabc_consumator_ContinueComponent_PhaseB', _cycle, outputDir=_output_dir, chosen_hue=_hue)
SingleSaveHistPlot(data_normal, 'Iabc_consumator_ContinueComponent_PhaseC', _cycle, outputDir=_output_dir, chosen_hue=_hue)
SingleSaveHistPlot(data_normal, 'Iabc_consumator_EffectiveComponent_PhaseA', _cycle, outputDir=_output_dir, chosen_hue=_hue)
SingleSaveHistPlot(data_normal, 'Iabc_consumator_EffectiveComponent_PhaseB', _cycle, outputDir=_output_dir, chosen_hue=_hue)
SingleSaveHistPlot(data_normal, 'Iabc_consumator_EffectiveComponent_PhaseC', _cycle, outputDir=_output_dir, chosen_hue=_hue)
SingleSaveHistPlot(data_normal, 'Iabc_consumator_FirstArmonic_PhaseA_Value', _cycle, outputDir=_output_dir, chosen_hue=_hue)
SingleSaveHistPlot(data_normal, 'Iabc_consumator_FirstArmonic_PhaseA_Angle', _cycle, outputDir=_output_dir, chosen_hue=_hue)
SingleSaveHistPlot(data_normal, 'Iabc_consumator_FirstArmonic_PhaseB_Value', _cycle, outputDir=_output_dir, chosen_hue=_hue)
SingleSaveHistPlot(data_normal, 'Iabc_consumator_FirstArmonic_PhaseB_Angle', _cycle, outputDir=_output_dir, chosen_hue=_hue)
SingleSaveHistPlot(data_normal, 'Iabc_consumator_FirstArmonic_PhaseC_Value', _cycle, outputDir=_output_dir, chosen_hue=_hue)
SingleSaveHistPlot(data_normal, 'Iabc_consumator_FirstArmonic_PhaseC_Angle', _cycle, outputDir=_output_dir, chosen_hue=_hue)
SingleSaveHistPlot(data_normal, 'Iabc_consumator_DeformingComponent_PhaseA', _cycle, outputDir=_output_dir, chosen_hue=_hue)
SingleSaveHistPlot(data_normal, 'Iabc_consumator_DeformingComponent_PhaseB', _cycle, outputDir=_output_dir, chosen_hue=_hue)
SingleSaveHistPlot(data_normal, 'Iabc_consumator_DeformingComponent_PhaseC', _cycle, outputDir=_output_dir, chosen_hue=_hue)
SingleSaveHistPlot(data_normal, 'Iabc_consumator_CoefficientDistorsion_PhaseA', _cycle, outputDir=_output_dir, chosen_hue=_hue)
SingleSaveHistPlot(data_normal, 'Iabc_consumator_CoefficientDistorsion_PhaseB', _cycle, outputDir=_output_dir, chosen_hue=_hue)
SingleSaveHistPlot(data_normal, 'Iabc_consumator_CoefficientDistorsion_PhaseC', _cycle, outputDir=_output_dir, chosen_hue=_hue)
SingleSaveHistPlot(data_normal, 'Iabc_consumator_CoefficientDistorsionContinue_PhaseA', _cycle, outputDir=_output_dir, chosen_hue=_hue)
SingleSaveHistPlot(data_normal, 'Iabc_consumator_CoefficientDistorsionContinue_PhaseB', _cycle, outputDir=_output_dir, chosen_hue=_hue)
SingleSaveHistPlot(data_normal, 'Iabc_consumator_CoefficientDistorsionContinue_PhaseC', _cycle, outputDir=_output_dir, chosen_hue=_hue)
SingleSaveHistPlot(data_normal, 'Iabc_consumator_DirectComponent_Value', _cycle, outputDir=_output_dir, chosen_hue=_hue)
SingleSaveHistPlot(data_normal, 'Iabc_consumator_DirectComponent_Angle', _cycle, outputDir=_output_dir, chosen_hue=_hue)
SingleSaveHistPlot(data_normal, 'Iabc_consumator_InverseComponent_Value', _cycle, outputDir=_output_dir, chosen_hue=_hue)
SingleSaveHistPlot(data_normal, 'Iabc_consumator_InverseComponent_Angle', _cycle, outputDir=_output_dir, chosen_hue=_hue)
SingleSaveHistPlot(data_normal, 'Iabc_consumator_ZeroComponent_Value', _cycle, outputDir=_output_dir, chosen_hue=_hue)
SingleSaveHistPlot(data_normal, 'Iabc_consumator_ZeroComponent_Angle', _cycle, outputDir=_output_dir, chosen_hue=_hue)
SingleSaveHistPlot(data_normal, 'Vabc_consumator_ContinueComponent_PhaseA', _cycle, outputDir=_output_dir, chosen_hue=_hue)
SingleSaveHistPlot(data_normal, 'Vabc_consumator_ContinueComponent_PhaseB', _cycle, outputDir=_output_dir, chosen_hue=_hue)
SingleSaveHistPlot(data_normal, 'Vabc_consumator_ContinueComponent_PhaseC', _cycle, outputDir=_output_dir, chosen_hue=_hue)
SingleSaveHistPlot(data_normal, 'Vabc_consumator_EffectiveComponent_PhaseA', _cycle, outputDir=_output_dir, chosen_hue=_hue)
SingleSaveHistPlot(data_normal, 'Vabc_consumator_EffectiveComponent_PhaseB', _cycle, outputDir=_output_dir, chosen_hue=_hue)
SingleSaveHistPlot(data_normal, 'Vabc_consumator_EffectiveComponent_PhaseC', _cycle, outputDir=_output_dir, chosen_hue=_hue)
SingleSaveHistPlot(data_normal, 'Vabc_consumator_FirstArmonic_PhaseA_Value', _cycle, outputDir=_output_dir, chosen_hue=_hue)
SingleSaveHistPlot(data_normal, 'Vabc_consumator_FirstArmonic_PhaseA_Angle', _cycle, outputDir=_output_dir, chosen_hue=_hue)
SingleSaveHistPlot(data_normal, 'Vabc_consumator_FirstArmonic_PhaseB_Value', _cycle, outputDir=_output_dir, chosen_hue=_hue)
SingleSaveHistPlot(data_normal, 'Vabc_consumator_FirstArmonic_PhaseB_Angle', _cycle, outputDir=_output_dir, chosen_hue=_hue)
SingleSaveHistPlot(data_normal, 'Vabc_consumator_FirstArmonic_PhaseC_Value', _cycle, outputDir=_output_dir, chosen_hue=_hue)
SingleSaveHistPlot(data_normal, 'Vabc_consumator_FirstArmonic_PhaseC_Angle', _cycle, outputDir=_output_dir, chosen_hue=_hue)
SingleSaveHistPlot(data_normal, 'Vabc_consumator_DeformingComponent_PhaseA', _cycle, outputDir=_output_dir, chosen_hue=_hue)
SingleSaveHistPlot(data_normal, 'Vabc_consumator_DeformingComponent_PhaseB', _cycle, outputDir=_output_dir, chosen_hue=_hue)
SingleSaveHistPlot(data_normal, 'Vabc_consumator_DeformingComponent_PhaseC', _cycle, outputDir=_output_dir, chosen_hue=_hue)
SingleSaveHistPlot(data_normal, 'Vabc_consumator_CoefficientDistorsion_PhaseA', _cycle, outputDir=_output_dir, chosen_hue=_hue)
SingleSaveHistPlot(data_normal, 'Vabc_consumator_CoefficientDistorsion_PhaseB', _cycle, outputDir=_output_dir, chosen_hue=_hue)
SingleSaveHistPlot(data_normal, 'Vabc_consumator_CoefficientDistorsion_PhaseC', _cycle, outputDir=_output_dir, chosen_hue=_hue)
SingleSaveHistPlot(data_normal, 'Vabc_consumator_CoefficientDistorsionContinue_PhaseA', _cycle, outputDir=_output_dir, chosen_hue=_hue)
SingleSaveHistPlot(data_normal, 'Vabc_consumator_CoefficientDistorsionContinue_PhaseB', _cycle, outputDir=_output_dir, chosen_hue=_hue)
SingleSaveHistPlot(data_normal, 'Vabc_consumator_CoefficientDistorsionContinue_PhaseC', _cycle, outputDir=_output_dir, chosen_hue=_hue)
SingleSaveHistPlot(data_normal, 'Vabc_consumator_DirectComponent_Value', _cycle, outputDir=_output_dir, chosen_hue=_hue)
SingleSaveHistPlot(data_normal, 'Vabc_consumator_DirectComponent_Angle', _cycle, outputDir=_output_dir, chosen_hue=_hue)
SingleSaveHistPlot(data_normal, 'Vabc_consumator_InverseComponent_Value', _cycle, outputDir=_output_dir, chosen_hue=_hue)
SingleSaveHistPlot(data_normal, 'Vabc_consumator_InverseComponent_Angle', _cycle, outputDir=_output_dir, chosen_hue=_hue)
SingleSaveHistPlot(data_normal, 'Vabc_consumator_ZeroComponent_Value', _cycle, outputDir=_output_dir, chosen_hue=_hue)
SingleSaveHistPlot(data_normal, 'Vabc_consumator_ZeroComponent_Angle', _cycle, outputDir=_output_dir, chosen_hue=_hue)


SingleSaveHistPlot_bins(data_normal, 'Iabc_generator_CoefficientDistorsionFirstArmonic_PhaseA', _cycle, outputDir=_output_dir, chosen_hue=_hue, _bins=c_bins)
SingleSaveHistPlot_bins(data_normal, 'Iabc_generator_CoefficientDistorsionFirstArmonic_PhaseB', _cycle, outputDir=_output_dir, chosen_hue=_hue, _bins=c_bins)
SingleSaveHistPlot_bins(data_normal, 'Iabc_generator_CoefficientDistorsionFirstArmonic_PhaseC', _cycle, outputDir=_output_dir, chosen_hue=_hue, _bins=c_bins)
SingleSaveHistPlot_bins(data_normal, 'Iabc_generator_CoefficeintNegativeAsimetry', _cycle, outputDir=_output_dir, chosen_hue=_hue, _bins=c_bins)
SingleSaveHistPlot_bins(data_normal, 'Iabc_generator_CoefficeintZeroAsimetry', _cycle, outputDir=_output_dir, chosen_hue=_hue, _bins=c_bins)
SingleSaveHistPlot_bins(data_normal, 'Vabc_generator_CoefficientDistorsionFirstArmonic_PhaseA', _cycle, outputDir=_output_dir, chosen_hue=_hue, _bins=c_bins)
SingleSaveHistPlot_bins(data_normal, 'Vabc_generator_CoefficientDistorsionFirstArmonic_PhaseB', _cycle, outputDir=_output_dir, chosen_hue=_hue, _bins=c_bins)
SingleSaveHistPlot_bins(data_normal, 'Vabc_generator_CoefficientDistorsionFirstArmonic_PhaseC', _cycle, outputDir=_output_dir, chosen_hue=_hue, _bins=c_bins)
SingleSaveHistPlot_bins(data_normal, 'Vabc_generator_CoefficeintNegativeAsimetry', _cycle, outputDir=_output_dir, chosen_hue=_hue, _bins=c_bins)
SingleSaveHistPlot_bins(data_normal, 'Vabc_generator_CoefficeintZeroAsimetry', _cycle, outputDir=_output_dir, chosen_hue=_hue, _bins=c_bins)
SingleSaveHistPlot_bins(data_normal, 'Iabc_loc_defect_CoefficientDistorsionFirstArmonic_PhaseA', _cycle, outputDir=_output_dir, chosen_hue=_hue, _bins=c_bins)
SingleSaveHistPlot_bins(data_normal, 'Iabc_loc_defect_CoefficientDistorsionFirstArmonic_PhaseB', _cycle, outputDir=_output_dir, chosen_hue=_hue, _bins=c_bins)
SingleSaveHistPlot_bins(data_normal, 'Iabc_loc_defect_CoefficientDistorsionFirstArmonic_PhaseC', _cycle, outputDir=_output_dir, chosen_hue=_hue, _bins=c_bins)
SingleSaveHistPlot_bins(data_normal, 'Iabc_loc_defect_CoefficeintNegativeAsimetry', _cycle, outputDir=_output_dir, chosen_hue=_hue, _bins=c_bins)
SingleSaveHistPlot_bins(data_normal, 'Iabc_loc_defect_CoefficeintZeroAsimetry', _cycle, outputDir=_output_dir, chosen_hue=_hue, _bins=c_bins)
SingleSaveHistPlot_bins(data_normal, 'Vabc_loc_defect_CoefficientDistorsionFirstArmonic_PhaseA', _cycle, outputDir=_output_dir, chosen_hue=_hue, _bins=c_bins)
SingleSaveHistPlot_bins(data_normal, 'Vabc_loc_defect_CoefficientDistorsionFirstArmonic_PhaseB', _cycle, outputDir=_output_dir, chosen_hue=_hue, _bins=c_bins)
SingleSaveHistPlot_bins(data_normal, 'Vabc_loc_defect_CoefficientDistorsionFirstArmonic_PhaseC', _cycle, outputDir=_output_dir, chosen_hue=_hue, _bins=c_bins)
SingleSaveHistPlot_bins(data_normal, 'Vabc_loc_defect_CoefficeintNegativeAsimetry', _cycle, outputDir=_output_dir, chosen_hue=_hue, _bins=c_bins)
SingleSaveHistPlot_bins(data_normal, 'Vabc_loc_defect_CoefficeintZeroAsimetry', _cycle, outputDir=_output_dir, chosen_hue=_hue, _bins=c_bins)
SingleSaveHistPlot_bins(data_normal, 'Iabc_consumator_CoefficientDistorsionFirstArmonic_PhaseA', _cycle, outputDir=_output_dir, chosen_hue=_hue, _bins=c_bins)
SingleSaveHistPlot_bins(data_normal, 'Iabc_consumator_CoefficientDistorsionFirstArmonic_PhaseB', _cycle, outputDir=_output_dir, chosen_hue=_hue, _bins=c_bins)
SingleSaveHistPlot_bins(data_normal, 'Iabc_consumator_CoefficientDistorsionFirstArmonic_PhaseC', _cycle, outputDir=_output_dir, chosen_hue=_hue, _bins=c_bins)
SingleSaveHistPlot_bins(data_normal, 'Iabc_consumator_CoefficeintNegativeAsimetry', _cycle, outputDir=_output_dir, chosen_hue=_hue, _bins=c_bins)
SingleSaveHistPlot_bins(data_normal, 'Iabc_consumator_CoefficeintZeroAsimetry', _cycle, outputDir=_output_dir, chosen_hue=_hue, _bins=c_bins)
SingleSaveHistPlot_bins(data_normal, 'Vabc_consumator_CoefficientDistorsionFirstArmonic_PhaseA', _cycle, outputDir=_output_dir, chosen_hue=_hue, _bins=c_bins)
SingleSaveHistPlot_bins(data_normal, 'Vabc_consumator_CoefficientDistorsionFirstArmonic_PhaseB', _cycle, outputDir=_output_dir, chosen_hue=_hue, _bins=c_bins)
SingleSaveHistPlot_bins(data_normal, 'Vabc_consumator_CoefficientDistorsionFirstArmonic_PhaseC', _cycle, outputDir=_output_dir, chosen_hue=_hue, _bins=c_bins)
SingleSaveHistPlot_bins(data_normal, 'Vabc_consumator_CoefficeintNegativeAsimetry', _cycle, outputDir=_output_dir, chosen_hue=_hue, _bins=c_bins)
SingleSaveHistPlot_bins(data_normal, 'Vabc_consumator_CoefficeintZeroAsimetry', _cycle, outputDir=_output_dir, chosen_hue=_hue, _bins=c_bins)
