==> 01_Compute and display date/05_ComputeDerivatedSignals.ipynb <==
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
%matplotlib inline
import seaborn as sns
import os

data_normal = pd.read_csv('Save_Normal_data.csv')
scaled_data_normal = pd.read_csv('Save_Normal_data_SCALED.csv')
scaled_fac = pd.read_csv('Normal_Scaled_Factor.csv')

listIdSimulare = list(data_normal["ID_Simulare"].unique())

dataColumn = list(data_normal.columns)[11:]
allColumn = list(data_normal.columns)

def derivateDataFrame_cycle(data_frame, all_columns, data_columns, List_Id_Simulare, _ignore_cycle = 0 ):
    final = list([all_columns])
    _data = data_frame
    for _id in List_Id_Simulare:
        aux = _data[_data["ID_Simulare"]==_id]
        for index in aux.index.to_list():
            if (aux.at[index, "cycle"] != _ignore_cycle):
                # not copying cycle 0
                result_row = list()
                for _col in all_columns:
                    if _col in data_columns:
                        _aux = (aux.at[index, _col] - aux.at[index - 1, _col])/(aux.at[index, "cycle"] - aux.at[index - 1, "cycle"])
                        result_row.append(_aux)
                    else:
                        result_row.append(aux.at[index, _col])
                final.append(result_row)
    return pd.DataFrame(final[1:], columns = final[0])

data_normal_dericate_1 = derivateDataFrame_cycle(data_normal, allColumn, dataColumn, listIdSimulare, 0)

scaled_data_normal_dericate_1 = derivateDataFrame_cycle(scaled_data_normal, allColumn, dataColumn, listIdSimulare, 0)

data_normal_dericate_2 = derivateDataFrame_cycle(data_normal_dericate_1, allColumn, dataColumn, listIdSimulare, 1)

scaled_data_normal_dericate_2 = derivateDataFrame_cycle(scaled_data_normal_dericate_1, allColumn, dataColumn, listIdSimulare, 1)

data_normal_dericate_1.to_csv('Save_Normal_data_Derivate1.csv')
data_normal_dericate_2.to_csv('Save_Normal_data_Derivate2.csv')

scaled_data_normal_dericate_1.to_csv('Save_Normal_data_SCALED_Derivate1.csv')
scaled_data_normal_dericate_2.to_csv('Save_Normal_data_SCALED_Derivate2.csv')
